### 2023-09-28 #5894 @editorialbot
[PRE REVIEW]: BGmisc: An R Package for Extended Behavior Genetics Analysis
**Submitting author:** @smasongarrison (<a href="http://orcid.org/0000-0002-4804-6003">S. Mason Garrison</a>)
**Repository:** https://github.com/R-Computing-Lab/BGmisc/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
