### 2023-09-28 #5894 @smasongarrison
> @editorialbot query scope
>
> @smasongarrison – this submission is one the smaller side of what we'd typically publish here so I'm going to start a scope query with the broader editorial team. This will take a couple of weeks to complete.

Could you expand upon this some more? Because I'm not sure, I'm understanding what you mean by small in scope.
