+++ R/validatePedigree.R
@@ -1,17 +1,89 @@
+#' Repair Pedigree
+#'
+#' This function applies a list of repair functions sequentially to a pedigree.
+#'
+#' @param ped A pedigree object.
+#' @param repair_funs A list of functions to repair the pedigree.
+#' @return A corrected pedigree.
+repairPedigree <- function(ped, repair_funs = NULL) {
+  corrected_ped <- ped
+  if(!is.null(repair_funs)){
+  for (fun in repair_funs) {
+    corrected_ped <- fun(corrected_ped)
+  }
+    return(corrected_ped)
+  } else if(is.null(repair_funs)){
+  corrected_ped <- repairIDs(corrected_ped)
+  corrected_ped <- repairParentIDs(corrected_ped)
+  
+  return(corrected_ped)
+} else { print ("You should never see this message. If you do, that means the repair_funs variable in repairPedigree is broken")
+}
+
+# Repairing IDs
+repairIDs <- function(ped) {
+  # Create a list to track the changes made
+  changes <- list()
+  # [logic to repair IDs]
+  return(ped)
+}
+
+# Repairing Parent IDs
+repairParentIDs <- function(ped) {
+  # [logic to repair Parent IDs]
+  return(ped)
+}
+
 #' Validate Pedigrees
 #'
-#' This function validates pedigrees based on several criteria
-
+#' This function validates pedigrees based on several criteria.
 #' @param ped A pedigree object
 #' @param verbose A logical indicating whether to print progress messages
 #' @param ... Additional arguments to be passed to \code{\link{validatePedigree}}
 #' @return A logical indicating whether the pedigree is valid and a list of warnings and ids of potentially invalid relationships
 #' @export
-#' 
+validatePedigree <- function(ped, verbose = FALSE) {
+  
+  corrected_ped <- repairPedigree(ped$ID, ped$dadID, ped$momID, ped$sex)
+  
+  # Validation checks
+  id_valid <- all(corrected_ped$ID == ped$ID)
+  dadID_valid <- all(corrected_ped$dadID == ped$dadID)
+  momID_valid <- all(corrected_ped$momID == ped$momID)
+  sex_valid <- all(corrected_ped$sex == ped$sex)
 
-validatePedigree() <- function(ped,
-                        verbose = FALSE){
+  # Compile results
+  is_valid <- id_valid && dadID_valid && momID_valid && sex_valid
+
+  # Prepare warnings and feedback
+  warnings <- list()
+  
+  if (!id_valid) {
+    warnings$id_warning <- "IDs in the corrected pedigree do not match the original IDs."
+  }
+  if (!dadID_valid) {
+    warnings$dadID_warning <- "Father IDs in the corrected pedigree do not match the original IDs."
+  }
+  if (!momID_valid) {
+    warnings$momID_warning <- "Mother IDs in the corrected pedigree do not match the original IDs."
+  }
+  if (!sex_valid) {
+    warnings$sex_warning <- "Sex values in the corrected pedigree do not match the original values."
+  }
+  
+  # Return results
+  if (verbose) {
+    return(list(is_valid = is_valid, corrected_ped = corrected_ped, warnings = warnings))
+  } else if (is_valid) {
+    return(corrected_ped)
+  } else {
+    print("Pedigree is not valid. Refer to the warnings for more details.")
+    return(warnings)
+  }
+}
 
 
-                            
-                        }
+
+
+
+
