+++ CONTRIBUTING.md
@@ -1,0 +1,77 @@
+# Contributing Guidelines for BGmisc
+
+Thank you for considering contributing to BGmisc. This document outlines the process and best practices for contributing to this R package hosted on GitHub under the R Computing Lab.
+## Table of Contents
+
+1. [Code of Conduct](#code-of-conduct)
+2. [Getting Started](#getting-started)
+3. [Bug Reports](#bug-reports)
+4. [Feature Requests](#feature-requests)
+5. [Pull Requests](#pull-requests)
+6. [Code Style](#code-style)
+7. [Testing](#testing)
+8. [Documentation](#documentation)
+9. [Communication](#communication)
+
+## Code of Conduct
+
+All contributors are expected to adhere to the project's [Code of Conduct](https://github.com/R-Computing-Lab/BGmisc/blob/main/CODE_OF_CONDUCT.md). Please read it carefully before contributing.
+
+## Getting Started
+
+1. Fork the [BGmisc repository](https://github.com/R-Computing-Lab/BGmisc) to your GitHub account.
+2. Clone the forked repository to your local machine.
+3. Install the required packages and set up the development environment.
+
+## Bug Reports
+
+For reporting bugs, please create an issue in the GitHub repository. Make sure to:
+
+- Provide a clear title and description.
+- Include a minimal reproducible example.
+- Tag the issue with the "bug" label.
+
+## Feature Requests
+
+New features are welcome. To request a new feature:
+
+- Open an issue in the GitHub repository.
+- Clearly describe the feature and its potential benefits.
+- Tag the issue with the "feature request" label.
+
+## Pull Requests
+
+1. Fork the repository and create a new branch for your work.
+2. Commit your changes in logical chunks.
+3. Open a pull request with a clear title and description.
+4. Make sure all existing tests pass.
+5. Add new tests for your changes.
+
+## Code Style
+
+Follow the [Tidyverse Style Guide](https://style.tidyverse.org/) for R programming to maintain code consistency.
+
+## Testing
+
+- Tests are implemented using the `testthat` package.
+- Make sure to add new tests for any added functionality.
+- Run all tests to ensure that they pass before submitting a pull request.
+
+## Documentation
+
+- Update the README.md and other relevant documentation.
+- Use [roxygen2](https://cran.r-project.org/web/packages/roxygen2/index.html) for documenting functions.
+- Include examples in the documentation where possible.
+
+## Communication
+
+- Use GitHub issues for all communication.
+- For more direct communication, you can [contact the maintainers](mailto:garrissm@wfu.edu).
+
+By contributing, you agree to abide by these guidelines and the project's [Code of Conduct](https://github.com/R-Computing-Lab/BGmisc/blob/main/CODE_OF_CONDUCT.md).
+
+Thank you for contributing to BGmisc!
+
+---
+
+Feel free to adapt these guidelines to better fit the specifics of the BGmisc project.
