+++ R/helper.R
@@ -133,55 +133,4 @@ relatedness <- function(...) {
   inferRelatedness(...)
 }
 
-#' Standardize Column Names in a Dataframe (Internal)
-#'
-#' This internal function standardizes the column names of a given dataframe.
-#' It is useful for ensuring that dataframes with varying column names can be processed by functions
-#' expecting specific column names.
-#'
-#' @param df A dataframe whose column names need to be standardized.
-#'
-#' @return A dataframe with standardized column names.
-#'
-#' @keywords internal
-standardize_colnames <- function(df) {
-  # Internal mapping of standardized names to possible variants
-  mapping <- list(
-    'fam' = c('fam', 'Fam', 'FAM',
-              'famID', 'FamID', 'FAMID',
-              'famid', 'Famid', 'FAMid',
-              'famfam','Famfam', 'FAMFAM'),
-    'ID' = c('ID', 'id', 'Id',
-             'Indiv', 'indiv', 'INDIV',
-             'individual', 'Individual', 'INDIVIDUAL'),
-    'gen' = c('gen', 'Gen', 'GEN',
-              'gens', 'Gens', 'GENS',
-              'generation', 'Generation', 'GENERATION'),
-    'dadID' = c('dadID', 'DadID', 'DADID',
-                'fatherID', 'FatherID', 'FATHERID'),
-    'momID' = c('momID', 'MomID', 'MOMID',
-                'motherID', 'MotherID', 'MOTHERID'),
-    'spt' = c('spt', 'Spt', 'SPT',
-                'spouse', 'Spouse', 'SPOUSE'),
-    'twinID' = c('twinID', 'twinid', 'TWINID',
-                'twin', 'Twin', 'TWIN'),
-    'sex' = c('sex', 'Sex', 'SEX',
-              'female', 'Female', 'FEMALE',
-              'gender', 'Gender', 'GENDER',
-              'male', 'Male', 'MALE',
-              'man', 'Man', 'MAN',
-              'men', 'Men', 'MEN',
-              'woman', 'Woman', 'WOMAN',
-              'women', 'Women', 'WOMEN')
-  )
-  for (standard_name in names(mapping)) {
-    for (variant in mapping[[standard_name]]) {
-      if (variant %in% colnames(df)) {
-        colnames(df)[colnames(df) == variant] <- standard_name
-        break  # Exit the loop once a match is found
-      }
-    }
-  }
 
-  return(df)
-}
