+++ R/validatePedigree.R
@@ -1,3 +1,77 @@
+#' Standardize Column Names in a Dataframe (Internal)
+#'
+#' This internal function standardizes the column names of a given dataframe.
+#' It is useful for ensuring that dataframes with varying column names can be processed by functions
+#' expecting specific column names.
+#'
+#' @param df A dataframe whose column names need to be standardized.
+#'
+#' @return A dataframe with standardized column names.
+#'
+#' @keywords internal
+standardize_colnames <- function(df) {
+  # Internal mapping of standardized names to possible variants
+  mapping <- list(
+    "fam" = c(
+      "fam", "Fam", "FAM",
+      "famID", "FamID", "FAMID",
+      "famid", "Famid", "FAMid",
+      "famfam", "Famfam", "FAMFAM"
+    ),
+    "ID" = c(
+      "ID", "id", "Id",
+      "Indiv", "indiv", "INDIV",
+      "individual", "Individual", "INDIVIDUAL"
+    ),
+    "gen" = c(
+      "gen", "Gen", "GEN",
+      "gens", "Gens", "GENS",
+      "generation", "Generation", "GENERATION"
+    ),
+    "dadID" = c(
+      "dadID", "DadID", "DADID",
+      "fatherID", "FatherID", "FATHERID"
+    ),
+    "momID" = c(
+      "momID", "MomID", "MOMID",
+      "motherID", "MotherID", "MOTHERID"
+    ),
+    "spt" = c(
+      "spt", "Spt", "SPT",
+      "sptID", "SptID", "SPTID",
+      "sptid", "Sptid", "SPTid",
+      "spouse", "Spouse", "SPOUSE",
+      "spouseID", "SpouseID", "SPOUSEID",
+      "spouseid", "Spouseid", "SPOUSEid",
+    ),
+    "twinID" = c(
+      "twinID", "twinid", "TWINID",
+      "twin", "Twin", "TWIN"
+    ),
+    "sex" = c(
+      "sex", "Sex", "SEX",
+      "female", "Female", "FEMALE",
+      "gender", "Gender", "GENDER",
+      "male", "Male", "MALE",
+      "man", "Man", "MAN",
+      "men", "Men", "MEN",
+      "woman", "Woman", "WOMAN",
+      "women", "Women", "WOMEN"
+    )
+  )
+  for (standard_name in names(mapping)) {
+    for (variant in mapping[[standard_name]]) {
+      if (variant %in% colnames(df)) {
+        colnames(df)[colnames(df) == variant] <- standard_name
+        break # Exit the loop once a match is found
+      }
+    }
+  }
+
+  return(df)
+}
+
+
 #' Repair Pedigree
 #'
 #' This function applies a list of repair functions sequentially to a pedigree.
@@ -6,18 +80,30 @@
 #' @param repair_funs A list of functions to repair the pedigree.
 #' @return A corrected pedigree.
 repairPedigree <- function(ped, repair_funs = NULL) {
-  corrected_ped <- ped
-  if(!is.null(repair_funs)){
-  for (fun in repair_funs) {
-    corrected_ped <- fun(corrected_ped)
-  }
+
+  corrected_ped <- ped <- standardize_colnames(ped)
+
+  # applies a list of repair functions sequentially to a pedigree.
+  if (!is.null(repair_funs)) {
+    for (fun in repair_funs) {
+      corrected_ped <- fun(corrected_ped)
+    }
     return(corrected_ped)
-  } else if(is.null(repair_funs)){
-  corrected_ped <- repairIDs(corrected_ped)
-  corrected_ped <- repairParentIDs(corrected_ped)
-  
-  return(corrected_ped)
-} else { print ("You should never see this message. If you do, that means the repair_funs variable in repairPedigree is broken")
[…]
