### 2023-10-02 457ac99
package revam
rename flatten_diag, convert renamer to use regex, fix verbose
