### 2023-10-02 ad2e36a
runs kinda...
the plot wrapper doesn't consistenly work

test
