+++ man/ped2mit.Rd
@@ -32,5 +32,5 @@ generations as there are in the data.}
 Take a pedigree and turn it into a mitochondrial relatedness matrix
 }
 \details{
-source examplePedigreeFunctions
+The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
 }
