+++ man/simulatePedigree.Rd
@@ -2,7 +2,12 @@
 % Please edit documentation in R/simulatePedigree.R
 \name{simulatePedigree}
 \alias{simulatePedigree}
-\title{Simulate Pedigrees}
+\title{Simulate Pedigrees
+This function simulates "balanced" pedigrees based on a group of parameters:
+1) k - Kids per couple;
+2) G - Number of generations;
+3) p - Proportion of males in offspring;
+4) r - Mating rate.}
 \usage{
 simulatePedigree(
   kpc = 3,
@@ -39,6 +44,7 @@ A \code{data.frame} with each row representing a simulated individual. The colum
 }
 }
 \description{
+Simulate Pedigrees
 This function simulates "balanced" pedigrees based on a group of parameters:
 1) k - Kids per couple;
 2) G - Number of generations;
