+++ man/PlotPedigree.Rd
@@ -8,7 +8,8 @@ A wrapped function to plot simulated pedigree from function \code{simulatePedigr
 plotPedigree(
   ped,
   code_male = NULL,
-  quiet = TRUE,
+  verbose = FALSE,
+  affected = NULL,
   cex = 0.5,
   col = 1,
   symbolsize = 1,
@@ -27,9 +28,16 @@ plotPedigree(
 \arguments{
 \item{ped}{The simulated pedigree data.frame from function \code{simulatePedigree}. Or a pedigree dataframe with the same colnames as the dataframe simulated from function \code{simulatePedigree}.}
 
-\item{code_male}{This optional input allows you to indicate what value in the sex variable codes for male.}
+\item{code_male}{This optional input allows you to indicate what value in the sex variable codes for male. Will be recoded as "M" (Male). If \code{NULL}, no recoding is performed.}
 
-\item{quiet}{Logical. If TRUE, suppresses the printing of the pedigree object. Default is TRUE.}
+\item{verbose}{logical  If TRUE, prints additional information. Default is FALSE.}
+
+\item{affected}{
+vector, or matrix with up to 4 columns for affected
+indicators. Subject's symbol is divided into sections for each
+status, shaded if indicator is 1, not-shaded for 0, and symbol "?"
+if missing (NA)
+}
 
 \item{cex}{The font size of the IDs for each individual in the plot.}
 
