+++ man/standardize_colnames.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helper.R
+% Please edit documentation in R/validatePedigree.R
 \name{standardize_colnames}
 \alias{standardize_colnames}
 \title{Standardize Column Names in a Dataframe (Internal)}
@@ -14,7 +14,8 @@ A dataframe with standardized column names.
 }
 \description{
 This internal function standardizes the column names of a given dataframe.
-It is useful for ensuring that dataframes with varying column names can be processed by functions
-expecting specific column names.
+It utilizes regular expressions and the `tolower()` function to match column names
+against a list of predefined standard names. The approach is case-insensitive and
+allows for flexible matching of column names.
 }
 \keyword{internal}
