+++ man/validatePedigree.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/validatePedigree.R
+\name{validatePedigree}
+\alias{validatePedigree}
+\title{Validate Pedigrees}
+\usage{
+validatePedigree(ped, verbose = FALSE)
+}
+\arguments{
+\item{ped}{A pedigree object}
+
+\item{verbose}{A logical indicating whether to print progress messages}
+
+\item{...}{Additional arguments to be passed to \code{\link{validatePedigree}}}
+}
+\value{
+A logical indicating whether the pedigree is valid and a list of warnings and ids of potentially invalid relationships
+}
+\description{
+This function validates pedigrees based on several criteria.
+}
