### 2023-10-04 d41a7b9
Update plotPedigree.R
