+++ R/convertPedigree.R
@@ -17,23 +17,43 @@ ped2com <- function(ped, component,
                     verbose = FALSE,
                     gc = FALSE,
                     flatten_diag = FALSE) {
-  component <- match.arg(tolower(component), choices = c("generation", "additive", "common nuclear", "mitochondrial"))
+
+  # Validate the 'component' argument and match it against predefined choices
+  component <- match.arg(tolower(component),
+    choices = c(
+      "generation",
+      "additive",
+      "common nuclear",
+      "mitochondrial"
+    )
+  )
+
+  # Get the number of rows in the pedigree dataset, representing the size of the family
   nr <- nrow(ped)
+
+  # Print the family size if verbose is TRUE
   if (verbose) {
     cat(paste0("Family Size = ", nr, "\n"))
   }
+  # Initialize variables
   parList <- list()
   lens <- integer(nr)
+
+  # Loop through each individual in the pedigree build the adjacency matrix for parent-child relationships
   # Is person in column j the parent of the person in row i? .5 for yes, 0 for no.
   for (i in 1:nr) {
     x <- ped[i, , drop = FALSE]
+
+    # Handle parentage according to the 'component' specified
     if (component %in% c("generation", "additive")) {
-      # iS mom of ID or iS dad of ID
+      # Code for 'generation' and 'additive' components
+      # Checks if is mom of ID or is dad of ID
       sMom <- (as.numeric(x["ID"]) == as.numeric(ped$momID))
       sDad <- (as.numeric(x["ID"]) == as.numeric(ped$dadID))
       val <- sMom | sDad
       val[is.na(val)] <- FALSE
     } else if (component %in% c("common nuclear")) {
+      # Code for 'common nuclear' component
       # IDs have the Same mom and Same dad
       sMom <- (as.numeric(x["momID"]) == as.numeric(ped$momID))
       sMom[is.na(sMom)] <- FALSE
@@ -41,6 +61,7 @@ ped2com <- function(ped, component,
       sDad[is.na(sDad)] <- FALSE
       val <- sMom & sDad
     } else if (component %in% c("mitochondrial")) {
+      # Code for 'mitochondrial' component
       sMom <- (as.numeric(x["ID"]) == as.numeric(ped$momID))
       sDad <- TRUE
       val <- sMom & sDad
@@ -48,22 +69,27 @@ ped2com <- function(ped, component,
     } else {
       stop("Unknown relatedness component requested")
     }
+    # Storing the indices of the parent-child relationships
     # keep track of indices only, and then initialize a single sparse matrix
     wv <- which(val)
     parList[[i]] <- wv
     lens[i] <- length(wv)
+    # Print progress if verbose is TRUE
     if (verbose && !(i %% 100)) {
       cat(paste0("Done with ", i, " of ", nr, "\n"))
     }
   }
+  # Construct sparse matrix
   jss <- rep(1L:nr, times = lens)
   iss <- unlist(parList)
+
+  # Garbage collection if gc is TRUE
   if (gc) {
     rm(parList, lens)
-  }
-  if (gc) {
     gc()
   }
+
+  # Set parent values depending on the component type
   if (component %in% c("generation", "additive")) {
     parVal <- .5
   } else if (component %in% c("common nuclear", "mitochondrial")) {
@@ -71,7 +97,10 @@ ped2com <- function(ped, component,
   } else {
     stop("Don't know how to set parental value")
   }
+
+  # Initialize adjacency matrix for parent-child relationships
   isPar <- Matrix::sparseMatrix(i = iss, j = jss, x = parVal, dims = c(nr, nr), dimnames = list(ped$ID, ped$ID))
+
   if (verbose) {
     cat("Completed first degree relatives (adjacency)\n")
   }
