+++ R/helper.R
@@ -132,5 +132,3 @@ relatedness <- function(...) {
   warning("The 'relatedness' function is deprecated. Please use 'inferRelatedness' instead.")
   inferRelatedness(...)
 }
-
-
