+++ R/evenInsert.R
@@ -3,6 +3,7 @@
 #'
 #' @param m A numeric vector of length less than or equal to n. The elements to be inserted.
 #' @param n A numeric vector. The vector into which the elements of m will be inserted.
+#' @param verbose logical  If TRUE, prints additional information. Default is FALSE.
 #' @return Returns a numeric vector with the elements of m evenly inserted into n.
 #' @details
 #' The function takes two vectors, m and n, and inserts the elements of m evenly into n.
@@ -11,7 +12,7 @@
 #' @export
 #' @seealso \code{\link{SimPed}} for the main function that uses this supporting function.
 
-evenInsert <- function(m, n) {
+evenInsert <- function(m, n, verbose = FALSE) {
   if (length(m) > length(n)) {
     temp <- m
     m <- n
@@ -22,11 +23,13 @@ evenInsert <- function(m, n) {
   for (i in 1:length(m)) {
     names(m)[i] <- ceiling(i * length(n) / length(m))
   }
-  # print(m)
-
+  if(verbose){
+   print(m)
+}
   names(n) <- 1:length(n)
-  # print(n)
-
+  if(verbose){
+  print(n)
+  }
   vec <- c(m, n)
   vec <- vec[order(as.numeric(names(vec)))]
   vec <- unname(vec)
