+++ R/validatePedigree.R
@@ -6,11 +6,11 @@
 #' allows for flexible matching of column names.
 #'
 #' @param df A dataframe whose column names need to be standardized.
-#'
+#' @param verbose A logical indicating whether to print progress messages.
 #' @return A dataframe with standardized column names.
 #'
 #' @keywords internal
-standardize_colnames <- function(df) {
+standardize_colnames <- function(df, verbose = FALSE) {
   # Internal mapping of standardized names to possible variants
   mapping <- list(
     "fam" = "^(?:fam(?:ily)?(?:id)?)",
@@ -22,7 +22,9 @@ standardize_colnames <- function(df) {
     "twinID" = "^(?:twin(?:id)?)",
     "sex" = "^(?:sex|gender|female|m(?:a(?:le|n)|en)|wom[ae]n)"
   )
-
+  if (verbose) {
+    print("Standardizing column names...")
+  }
   lowered_colnames <- tolower(colnames(df))
   for (standard_name in names(mapping)) {
     regex_pattern <- mapping[[standard_name]]
@@ -48,7 +50,6 @@ standardize_colnames <- function(df) {
 #' @seealso \code{\link{plotPedigree}}
 #'
 recode_sex <- function(ped, code_male) {
-
   # Recode as "F" or "M" based on code_male, preserving NAs
   if (!is.null(code_male)) {
     # Initialize sex_recode as NA, preserving the length of the 'sex' column
@@ -71,7 +72,6 @@ recode_sex <- function(ped, code_male) {
 #' @param repair_funs A list of functions to repair the pedigree.
 #' @return A corrected pedigree.
 repairPedigree <- function(ped, repair_funs = NULL) {
-
   corrected_ped <- ped <- standardize_colnames(ped)
 
   # applies a list of repair functions sequentially to a pedigree.
@@ -80,9 +80,10 @@ repairPedigree <- function(ped, repair_funs = NULL) {
       corrected_ped <- fun(corrected_ped)
     }
     return(corrected_ped)
-  # if not provided, use the default repair functions
+    # if not provided, use the default repair functions
   } else if (is.null(repair_funs)) {
     corrected_ped <- repairIDs(corrected_ped)
+    corrected_ped <- repairSex(corrected_ped)
     corrected_ped <- repairParentIDs(corrected_ped)
 
     return(corrected_ped)
@@ -92,62 +93,60 @@ repairPedigree <- function(ped, repair_funs = NULL) {
 }
 
 # To do
-# - Missing parents: If one parent is missing and the other one isn't, this needs to be handled somehow. Firstly, I think it can cause certain ways of estimating relatedness to give wrong numbers. And secondly, it requires us to make guesses in cases where e.g. two people have the same mother and missing fathers: They could then be either half-sibs, if the missing fathers are different people, or full sibs if not. This then also affects relatedness for their descendants.
 # - Missing rows: Sometimes, ID-codes in the mother or father column do not exist in the ID-column. That is, the people listed as someone's parents sometimes do not have their own rows, with columns for their parents etc.
 # - Wrong IDs: It is possible that the ID code written for e.g. someone's mother is simply written wrong. This is especially problematic if there are people in the file who actually have the code that was mistakenly given.
 # - A person's child being registered as their parent: I randomly found a case of this. A girl was registered as her father's mother.
[…]
