+++ man/repairParentIDs.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/checkParents.R
+\name{repairParentIDs}
+\alias{repairParentIDs}
+\title{Repair Parent IDs}
+\usage{
+repairParentIDs(ped, verbose = FALSE)
+}
+\arguments{
+\item{ped}{A pedigree object}
+
+\item{verbose}{A logical indicating whether to print progress messages}
+}
+\value{
+A corrected pedigree
+}
+\description{
+This function repairs parent IDs in a pedigree.
+}

[…]
