### 2023-10-05 bf95371
Update README.Rmd
[![Dev branch](https://github.com/R-Computing-Lab/BGmisc/actions/workflows/R-CMD-devcheck.yaml/badge.svg)](https://github.com/R-Computing-Lab/BGmisc/actions/workflows/R-CMD-devcheck.yaml)
