### 2023-10-05 c26b023
Better but still have the bug of detection failure
