+++ man/recode_sex.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/validatePedigree.R
-\name{recode_sex}
-\alias{recode_sex}
-\title{recode_sex
+\name{recodeSex}
+\alias{recodeSex}
+\title{recodeSex
 An internal utility function for \code{plotPedigree} to handle the recoding of the 'sex' variable in a pedigree dataframe.}
 \usage{
-recode_sex(ped, code_male)
+recodeSex(ped, code_male)
 }
 \arguments{
 \item{ped}{A pedigree data.frame containing the relevant columns, including 'sex'.}
@@ -16,7 +16,7 @@ recode_sex(ped, code_male)
 A modified version of the input data.frame \code{ped}, containing an additional or modified 'sex_recode' column where the 'sex' values are recoded according to \code{code_male}. NA values in the 'sex' column are preserved.
 }
 \description{
-recode_sex
+recodeSex
 An internal utility function for \code{plotPedigree} to handle the recoding of the 'sex' variable in a pedigree dataframe.
 }
 \seealso{
