+++ R/famSizeCal.R
@@ -69,14 +69,14 @@ famSizeCal <- function(kpc, Ngen, marR) {
     )
     size <- sum(allGens)
   } else {
-    stop("You should never see this message. 
+    stop("You should never see this message.
     If you do, that means that famSizeCal is not working properly.")
   }
   return(size)
 }
 
 #' twinImpute
-#' A function to impute twins in the simulated pedigree \code{data.frame}. 
+#' A function to impute twins in the simulated pedigree \code{data.frame}.
 #' Twins can be imputed by specifying their IDs or by specifying the generation the twin should be imputed.
 #' This is a supplementary function for \code{simulatePedigree}.
 #' @param ped A \code{data.frame} in the same format as the output of \code{simulatePedigree}.
@@ -86,47 +86,47 @@ famSizeCal <- function(kpc, Ngen, marR) {
 #' @return Returns a \code{data.frame} with MZ twins infomration added as a new column.
 #' @export
 
-# A function to impute twins in the simulated pedigree \code{data.frame}. 
+# A function to impute twins in the simulated pedigree \code{data.frame}.
 # Twins can be imputed by specifying their IDs or by specifying the generation the twin should be imputed.
-twinImpute <- function(ped, ID_twin1 = NA_integer_, ID_twin2 = NA_integer_, gen_twin = 2){
+twinImpute <- function(ped, ID_twin1 = NA_integer_, ID_twin2 = NA_integer_, gen_twin = 2) {
   # a support function
   resample <- function(x, ...) x[sample.int(length(x), ...)]
   # Check if the ped is the same format as the output of simulatePedigree
-  if(paste0(colnames(ped), collapse = "") != paste0(c("fam", "ID", "gen", "dadID", "momID", "spt", "sex"), collapse = "")){
+  if (paste0(colnames(ped), collapse = "") != paste0(c("fam", "ID", "gen", "dadID", "momID", "spt", "sex"), collapse = "")) {
     stop("The input pedigree is not in the same format as the output of simulatePedigree")
   }
   ped$MZtwin <- NA_integer_
   # Check if the two IDs are provided
-  if(is.na(ID_twin1) || is.na(ID_twin2)){
+  if (is.na(ID_twin1) || is.na(ID_twin2)) {
     # Check if the generation is provided
-    if(is.na(gen_twin)){
+    if (is.na(gen_twin)) {
       stop("You should provide either the IDs of the twins or the generation of the twins")
     } else {
       # Check if the generation is valid
-      if(gen_twin < 2 || gen_twin > max(ped$gen)){
+      if (gen_twin < 2 || gen_twin > max(ped$gen)) {
         stop("The generation of the twins should be an integer between 2 and the maximum generation in the pedigree")
       } else {
-        idx <- nrow(ped[ped$gen == gen_twin & !is.na(ped$dadID),]) 
-        usedID = c()
+        idx <- nrow(ped[ped$gen == gen_twin & !is.na(ped$dadID), ])
+        usedID <- c()
         # randomly loop through all the indivuduals in the generation until find an individual who is the same sex and shares the same dadID and momID with another individual
-        for (i in 1: idx) {
-          cat("loop",i)
+        for (i in 1:idx) {
+          cat("loop", i)
           # check if i is equal to the number of individuals in the generation
           usedID <- c(usedID, ID_twin1)
-          #print(usedID)
-          if(i < idx){
+          # print(usedID)
+          if (i < idx) {
[…]
