### 2023-10-07 a02eec5
Update convertPedigree.R
adding alias because I forget the i a lot
