### 2023-10-09 0d60c3b
moved recoded sex to own variable
