### 2023-10-10 7d589ec
Update checkParents.R
