+++ R/simulatePedigree.R
@@ -11,7 +11,7 @@
 #' @param sexR Sex ratio of offspring. A numeric value ranging from 0 to 1 that determines the proportion of males in all offspring in this pedigree. For instance, 0.4 means 40 percent of the offspring will be male.
 #' @param marR Mating rate. A numeric value ranging from 0 to 1 which determines the proportion of mated (fertilized) couples in the pedigree within each generation. For instance, marR = 0.5 suggests 50 percent of the offspring in a specific generation will be mated and have their offspring.
 #' @param balancedSex Not fully developed yet. Always \code{TRUE} in the current version.
-#' @param balancedmar Not fully developed yet. Always \code{TRUE} in the current version.
+#' @param balancedMar Not fully developed yet. Always \code{TRUE} in the current version.
 #' @param verbose logical  If TRUE, print progress through stages of algorithm
 
 #' @return A \code{data.frame} with each row representing a simulated individual. The columns are as follows:
@@ -31,7 +31,7 @@ simulatePedigree <- function(kpc = 3,
                              sexR = .5,
                              marR = 2 / 3,
                              balancedSex = TRUE,
-                             balancedmar = TRUE,
+                             balancedMar = TRUE,
                              verbose = FALSE) {
   # SexRatio: ratio of male over female in the offspring setting; used in the between generation combinations
   SexRatio <- sexR / (1 - sexR)
