+++ NAMESPACE
@@ -12,6 +12,7 @@ export(famSizeCal)
 export(fitComponentModel)
 export(identifyComponentModel)
 export(inferRelatedness)
+export(makeTwins)
 export(ped2add)
 export(ped2ce)
 export(ped2cn)
@@ -24,7 +25,6 @@ export(related_coef)
 export(relatedness)
 export(repairSex)
 export(simulatePedigree)
-export(twinImpute)
 export(validatePedigree)
 export(vech)
 import(kinship2)
