+++ R/tweakPedigree.R
@@ -1,0 +1,156 @@
+#' makeTwins
+#' A function to impute twins in the simulated pedigree \code{data.frame}.
+#' Twins can be imputed by specifying their IDs or by specifying the generation the twin should be imputed.
+#' This is a supplementary function for \code{simulatePedigree}.
+#' @param ped A \code{data.frame} in the same format as the output of \code{simulatePedigree}.
+#' @param ID_twin1 A vector of \code{ID} of the first twin.
+#' @param ID_twin2 A vector of \code{ID} of the second twin.
+#' @param verbose logical.  If TRUE, print progress through stages of algorithm
+#' @param gen_twin A vector of \code{generation} of the twin to be imputed.
+#' @return Returns a \code{data.frame} with MZ twins infomration added as a new column.
+#' @export
+
+# A function to impute twins in the simulated pedigree \code{data.frame}.
+# Twins can be imputed by specifying their IDs or by specifying the generation the twin should be imputed.
+makeTwins <- function(ped, ID_twin1 = NA_integer_, ID_twin2 = NA_integer_, gen_twin = 2, verbose = FALSE) {
+  # a support function
+  resample <- function(x, ...) x[sample.int(length(x), ...)]
+  # Check if the ped is the same format as the output of simulatePedigree
+  if (paste0(colnames(ped), collapse = "") != paste0(c("fam", "ID", "gen",
+   "dadID", "momID", "spt", "sex"), collapse = "")) {
+    ped <- standardize_colnames(ped)
+    if(verbose){
+      cat("The input pedigree is not in the same format as the output of simulatePedigree\n")
+    }
+    #stop("The input pedigree is not in the same format as the output of simulatePedigree")
+  }
+  ped$MZtwin <- NA_integer_
+  # Check if the two IDs are provided
+  if (is.na(ID_twin1) || is.na(ID_twin2)) {
+    # Check if the generation is provided
+    if (is.na(gen_twin)) {
+      stop("You should provide either the IDs of the twins or the generation of the twins")
+    } else {
+      # Check if the generation is valid
+      if (gen_twin < 2 || gen_twin > max(ped$gen)) {
+        stop("The generation of the twins should be an integer between 2 and the maximum generation in the pedigree")
+      } else {
+        idx <- nrow(ped[ped$gen == gen_twin & !is.na(ped$dadID), ])
+        usedID <- c()
+        # randomly loop through all the indivuduals in the generation until find an individual who is the same sex and shares the same dadID and momID with another individual
+        for (i in 1:idx) {
+          cat("loop", i)
+          # check if i is equal to the number of individuals in the generation
+          usedID <- c(usedID, ID_twin1)
+          # print(usedID)
+          if (i < idx) {
+            # randomly select one individual from the generation
+            ID_twin1 <- resample(ped$ID[ped$gen == gen_twin & !(ped$ID %in% usedID) & !is.na(ped$dadID)], 1)
+            # cat("twin1", ID_twin1, "\n")
+            # find one same sex sibling who has the same dadID and momID as the selected individual
+            twin2_Pool <- ped$ID[ped$ID != ID_twin1 & ped$gen == gen_twin & ped$sex == ped$sex[ped$ID == ID_twin1] & ped$dadID == ped$dadID[ped$ID == ID_twin1] & ped$momID == ped$momID[ped$ID == ID_twin1]]
+            # if there is an non-NA value in the twin2_Pool, get rid of the NA value
+            if (all(is.na(twin2_Pool))) {
+              cat("twin2_Pool is all NA\n")
+              next
+            } else {
+              twin2_Pool <- twin2_Pool[!is.na(twin2_Pool)]
[…]
