+++ R/convertPedigree.R
@@ -100,11 +100,13 @@ ped2com <- function(ped, component,
   }
 
   # Initialize adjacency matrix for parent-child relationships
-  isPar <- Matrix::sparseMatrix(i = iss,
-                                j = jss,
-                                x = parVal,
-                                dims = c(nr, nr),
-                                dimnames = list(ped$ID, ped$ID))
+  isPar <- Matrix::sparseMatrix(
+    i = iss,
+    j = jss,
+    x = parVal,
+    dims = c(nr, nr),
+    dimnames = list(ped$ID, ped$ID)
+  )
 
   if (verbose) {
     cat("Completed first degree relatives (adjacency)\n")
@@ -202,7 +204,7 @@ ped2add <- function(ped, max.gen = Inf, sparse = FALSE, verbose = FALSE, gc = FA
 #' @inheritParams ped2com
 #' @details The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
 #' @export
-#'@aliases ped2mt
+#' @aliases ped2mt
 #'
 ped2mit <- ped2mt <- function(ped, max.gen = Inf, sparse = FALSE, verbose = FALSE, gc = FALSE, flatten.diag = FALSE) {
   ped2com(
