+++ man/repairPedigree.Rd
@@ -4,12 +4,27 @@
 \alias{repairPedigree}
 \title{Repair Pedigree}
 \usage{
-repairPedigree(ped, repair_funs = NULL, ...)
+repairPedigree(
+  ped,
+  repair_funs = NULL,
+  check_id = TRUE,
+  check_sex = TRUE,
+  check_parents = TRUE,
+  verbose = FALSE
+)
 }
 \arguments{
 \item{ped}{A pedigree object.}
 
 \item{repair_funs}{A list of functions to repair the pedigree.}
+
+\item{check_id}{Logical. Indicates that IDs should be validated}
+
+\item{check_sex}{Logical. Indicates that sex should be validated}
+
+\item{check_parents}{Logical. Indicates that parents should be validated}
+
+\item{verbose}{Logical. Indicates whether to print progress messages}
 }
 \value{
 A corrected pedigree.
