+++ R/convertPedigree.R
@@ -5,11 +5,10 @@
 #'  (e.g., only up to 4th degree relatives).  The default of Inf uses as many
 #'  generations as there are in the data.
 #' @param sparse logical.  If TRUE, use and return sparse matrices from Matrix package
-#' @param verbose logical.  If TRUE, print progress through stages of algorithm
+#' @param verbose logical  If TRUE, print progress through stages of algorithm
 #' @param gc logical. If TRUE, do frequent garbage collection via \code{\link{gc}} to save memory
-#' @param flatten.diag logical. If TRUE, overwrite the diagonal of the final relatedness matrix with ones
-#' @param ... additional arguments to be passed to \code{\link{ped2com}}
-#' @details The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
+#' @param flatten.diag Logical. The default is FALSE. If TRUE, overwrites the diagonal of the final relatedness matrix with ones.
+#' @details source examplePedigreeFunctions
 #' @export
 #'
 ped2com <- function(ped, component,
@@ -17,44 +16,24 @@ ped2com <- function(ped, component,
                     sparse = FALSE,
                     verbose = FALSE,
                     gc = FALSE,
-                    flatten.diag = FALSE,
-                    ...) {
-  # Validate the 'component' argument and match it against predefined choices
-  component <- match.arg(tolower(component),
-    choices = c(
-      "generation",
-      "additive",
-      "common nuclear",
-      "mitochondrial"
-    )
-  )
-
-  # Get the number of rows in the pedigree dataset, representing the size of the family
+                    flatten.diag = FALSE) {
+  component <- match.arg(tolower(component), choices = c("generation", "additive", "common nuclear", "mitochondrial"))
   nr <- nrow(ped)
-
-  # Print the family size if verbose is TRUE
   if (verbose) {
     cat(paste0("Family Size = ", nr, "\n"))
   }
-  # Initialize variables
   parList <- list()
   lens <- integer(nr)
-
-  # Loop through each individual in the pedigree build the adjacency matrix for parent-child relationships
   # Is person in column j the parent of the person in row i? .5 for yes, 0 for no.
   for (i in 1:nr) {
     x <- ped[i, , drop = FALSE]
-
-    # Handle parentage according to the 'component' specified
     if (component %in% c("generation", "additive")) {
-      # Code for 'generation' and 'additive' components
-      # Checks if is mom of ID or is dad of ID
+      # iS mom of ID or iS dad of ID
       sMom <- (as.numeric(x["ID"]) == as.numeric(ped$momID))
       sDad <- (as.numeric(x["ID"]) == as.numeric(ped$dadID))
       val <- sMom | sDad
       val[is.na(val)] <- FALSE
     } else if (component %in% c("common nuclear")) {
-      # Code for 'common nuclear' component
       # IDs have the Same mom and Same dad
       sMom <- (as.numeric(x["momID"]) == as.numeric(ped$momID))
       sMom[is.na(sMom)] <- FALSE
@@ -62,7 +41,6 @@ ped2com <- function(ped, component,
       sDad[is.na(sDad)] <- FALSE
       val <- sMom & sDad
     } else if (component %in% c("mitochondrial")) {
-      # Code for 'mitochondrial' component
       sMom <- (as.numeric(x["ID"]) == as.numeric(ped$momID))
       sDad <- TRUE
       val <- sMom & sDad
@@ -70,27 +48,22 @@ ped2com <- function(ped, component,
[…]
