+++ R/evenInsert.R
@@ -3,7 +3,6 @@
 #'
 #' @param m A numeric vector of length less than or equal to n. The elements to be inserted.
 #' @param n A numeric vector. The vector into which the elements of m will be inserted.
-#' @param verbose logical  If TRUE, prints additional information. Default is FALSE.
 #' @return Returns a numeric vector with the elements of m evenly inserted into n.
 #' @details
 #' The function takes two vectors, m and n, and inserts the elements of m evenly into n.
@@ -12,7 +11,7 @@
 #' @export
 #' @seealso \code{\link{SimPed}} for the main function that uses this supporting function.
 
-evenInsert <- function(m, n, verbose = FALSE) {
+evenInsert <- function(m, n) {
   if (length(m) > length(n)) {
     temp <- m
     m <- n
@@ -23,13 +22,11 @@ evenInsert <- function(m, n, verbose = FALSE) {
   for (i in 1:length(m)) {
     names(m)[i] <- ceiling(i * length(n) / length(m))
   }
-  if (verbose) {
-    print(m)
-  }
+  # print(m)
+
   names(n) <- 1:length(n)
-  if (verbose) {
-    print(n)
-  }
+  # print(n)
+
   vec <- c(m, n)
   vec <- vec[order(as.numeric(names(vec)))]
   vec <- unname(vec)
