+++ R/validatePedigree.R
@@ -40,50 +40,50 @@ standardize_colnames <- function(df, verbose = FALSE) {
 }
 
 
-#' Repair Pedigree
-#'
-#' This function applies a list of repair functions sequentially to a pedigree.
-#'
-#' @param ped A pedigree object.
-#' @param repair_funs A list of functions to repair the pedigree.
-#' @param verbose Logical. Indicates whether to print progress messages
-#' @param check_sex Logical. Indicates that sex should be validated
-#' @param check_parents Logical. Indicates that parents should be validated
-#' @param check_id Logical. Indicates that IDs should be validated
-#' @return A corrected pedigree.
-repairPedigree <- function(
-    ped,
-    repair_funs = NULL,
-    check_id = TRUE,
-    check_sex = TRUE,
-    check_parents = TRUE,
-    verbose = FALSE) {
-  corrected_ped <- ped <- standardize_colnames(ped)
-  if (verbose) {
-    print("Repairing pedigree...")
-  }
-  # applies a list of repair functions sequentially to a pedigree.
-  if (!is.null(repair_funs)) {
-    for (fun in repair_funs) {
-      corrected_ped <- fun(corrected_ped)
-    }
-    return(corrected_ped)
-    # if not provided, use the default repair functions
-  } else if (is.null(repair_funs)) {
-    if (check_id) {
-      corrected_ped <- repairIDs(corrected_ped)
-    }
-    if (check_sex) {
-      corrected_ped <- repairSex(corrected_ped)
-    }
-    if (check_parents) {
-      corrected_ped <- repairParentIDs(corrected_ped)
-    }
-    return(corrected_ped)
-  } else {
-    print("You should never see this message. If you do, that means the repair_funs variable in repairPedigree is broken")
-  }
-}
+# Repair Pedigree
+#
+# This function applies a list of repair functions sequentially to a pedigree.
+#
+# @param ped A pedigree object.
+# @param repair_funs A list of functions to repair the pedigree.
+# @param verbose Logical. Indicates whether to print progress messages
+# @param check_sex Logical. Indicates that sex should be validated
+# @param check_parents Logical. Indicates that parents should be validated
+# @param check_id Logical. Indicates that IDs should be validated
+# @return A corrected pedigree.
+# repairPedigree <- function(
+    # ped,
+    # repair_funs = NULL,
+    # check_id = TRUE,
+    # check_sex = TRUE,
+    # check_parents = TRUE,
+    # verbose = FALSE) {
+  # corrected_ped <- ped <- standardize_colnames(ped)
+  # if (verbose) {
+    # print("Repairing pedigree...")
+  # }
+  # # applies a list of repair functions sequentially to a pedigree.
+  # if (!is.null(repair_funs)) {
+    # for (fun in repair_funs) {
+      # corrected_ped <- fun(corrected_ped)
+    # }
+    # return(corrected_ped)
+    # # if not provided, use the default repair functions
+  # } else if (is.null(repair_funs)) {
+    # if (check_id) {
+      # corrected_ped <- repairIDs(corrected_ped)
+    # }
+    # if (check_sex) {
+      # corrected_ped <- repairSex(corrected_ped)
+    # }
+    # if (check_parents) {
+      # corrected_ped <- repairParentIDs(corrected_ped)
+    # }
+    # return(corrected_ped)
+  # } else {
+    # print("You should never see this message. If you do, that means the repair_funs variable in repairPedigree is broken")
+  # }
+# }
 
 # To do
 # - Missing rows: Sometimes, ID-codes in the mother or father column do not exist in the ID-column. That is, the people listed as someone's parents sometimes do not have their own rows, with columns for their parents etc.
[…]
