+++ NAMESPACE
@@ -4,7 +4,6 @@ export(SimPed)
 export(allGens)
 export(calculateRelatedness)
 export(checkIDs)
-export(checkParentIDs)
 export(checkSex)
 export(comp2vech)
 export(evenInsert)
@@ -25,7 +24,6 @@ export(related_coef)
 export(relatedness)
 export(repairSex)
 export(simulatePedigree)
-export(validatePedigree)
 export(vech)
 import(kinship2)
 importFrom(stats,runif)
