+++ man/ped2add.Rd
@@ -22,15 +22,16 @@ generations as there are in the data.}
 
 \item{sparse}{logical.  If TRUE, use and return sparse matrices from Matrix package}
 
-\item{verbose}{logical  If TRUE, print progress through stages of algorithm}
+\item{verbose}{logical.  If TRUE, print progress through stages of algorithm}
 
 \item{gc}{logical. If TRUE, do frequent garbage collection via \code{\link{gc}} to save memory}
 
-\item{flatten.diag}{Logical. The default is FALSE. If TRUE, overwrites the diagonal of the final relatedness matrix with ones.}
+\item{flatten.diag}{logical. If TRUE, overwrite the diagonal of the final relatedness matrix with ones}
 }
 \description{
 Take a pedigree and turn it into an additive genetics relatedness matrix
 }
 \details{
-source examplePedigreeFunctions
+The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
+For more advanced scenarios and detailed explanations, consult this vignette.
 }
