### 2023-10-13 8bbedca
Create validatePedigree.R
