+++ R/validatePedigree.R
@@ -1,0 +1,175 @@
+#' Standardize Column Names in a Dataframe (Internal)
+#'
+#' This internal function standardizes the column names of a given dataframe.
+#' It utilizes regular expressions and the `tolower()` function to match column names
+#' against a list of predefined standard names. The approach is case-insensitive and
+#' allows for flexible matching of column names.
+#'
+#' @param df A dataframe whose column names need to be standardized.
+#' @param verbose A logical indicating whether to print progress messages.
+#' @return A dataframe with standardized column names.
+#'
+#' @keywords internal
+standardize_colnames <- function(df, verbose = FALSE) {
+  # Internal mapping of standardized names to possible variants
+  mapping <- list(
+    "fam" = "^(?:fam(?:ily)?(?:id)?)",
+    "ID" = "^(?:i(?:d$|ndiv(?:idual)?))",
+    "gen" = "^(?:gen(?:s|eration)?)",
+    "dadID" = "^(?:d(?:ad)?id|fatherid)",
+    "momID" = "^(?:m(?:om)?id|motherid)",
+    "spt" = "^(?:s(?:pt)?id|spouse(?:id)?)",
+    "twinID" = "^(?:twin(?:id)?)",
+    "sex" = "^(?:sex|gender|female|m(?:a(?:le|n)|en)|wom[ae]n)"
+  )
+  if (verbose) {
+    print("Standardizing column names...")
+  }
+  lowered_colnames <- tolower(colnames(df))
+  for (standard_name in names(mapping)) {
+    regex_pattern <- mapping[[standard_name]]
+    matched_variant <- grep(regex_pattern, lowered_colnames, value = TRUE)
+    if (length(matched_variant) > 0) {
+      # Update the first match in original case
+      original_matched <- colnames(df)[tolower(colnames(df)) == matched_variant[1]]
+      colnames(df)[colnames(df) == original_matched] <- standard_name
+    }
+  }
+
+  return(df)
+}
+
+
+# Repair Pedigree
+#
+# This function applies a list of repair functions sequentially to a pedigree.
+#
+# @param ped A pedigree object.
+# @param repair_funs A list of functions to repair the pedigree.
+# @param verbose Logical. Indicates whether to print progress messages
+# @param check_sex Logical. Indicates that sex should be validated
+# @param check_parents Logical. Indicates that parents should be validated
+# @param check_id Logical. Indicates that IDs should be validated
+# @return A corrected pedigree.
+# repairPedigree <- function(
+    # ped,
+    # repair_funs = NULL,
+    # check_id = TRUE,
+    # check_sex = TRUE,
+    # check_parents = TRUE,
+    # verbose = FALSE) {
+  # corrected_ped <- ped <- standardize_colnames(ped)
+  # if (verbose) {
+    # print("Repairing pedigree...")
+  # }
+  # # applies a list of repair functions sequentially to a pedigree.
+  # if (!is.null(repair_funs)) {
+    # for (fun in repair_funs) {
+      # corrected_ped <- fun(corrected_ped)
+    # }
+    # return(corrected_ped)
+    # # if not provided, use the default repair functions
+  # } else if (is.null(repair_funs)) {
+    # if (check_id) {
+      # corrected_ped <- repairIDs(corrected_ped)
+    # }
+    # if (check_sex) {
+      # corrected_ped <- repairSex(corrected_ped)
+    # }
+    # if (check_parents) {
+      # corrected_ped <- repairParentIDs(corrected_ped)
+    # }
+    # return(corrected_ped)
+  # } else {
+    # print("You should never see this message. If you do, that means the repair_funs variable in repairPedigree is broken")
+  # }
+# }
+
+# To do
[…]
