### 2023-10-13 0cbc297
Update checkSex.R
