### 2023-10-27 47436a6
Use sparse assignment for mitochondrial nonzero elements
