### 2023-10-27 1a82c18
Add potter dataset with documentation
