### 2023-10-27 87b0102
Fix whitespace in inbreeding data man page
