### 2023-10-27 13eaa61
Move verbose comment to correct location
