+++ data-raw/HP_df.R
@@ -1,55 +1,44 @@
----
-title: "Harry Potter Pedigree"
-author: "Rachel Good"
-date: "2023-10-18"
-output: html_document
----
 
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
 #devtools::install_github("R-Computing-Lab/BGmisc")
-library(BGmisc)
 library(tidyverse)
-```
 
 
 ## Create dataframe
 
-```{r createdf}
 hp_ped <- data.frame(
-  ID = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 
-         11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 
+  ID = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
+         11, 12, 13, 14, 15, 16, 17, 18, 19, 20,
          21, 22, 23, 24, 25, 26, 27, 28, 29, 30),
   FamID = c("fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1"),
-  name = c("Vernon Dursley", 
-           "Marjorie Dursley", 
-           "Petunia Evans", 
-           "Lily Evans", 
-           "James Potter", 
-           "Dudley Dursley", 
-           "Harry Potter", 
-           "Ginny Weasley", 
-           "Arthur Weasley", 
-           "Molly Prewett", 
-           "Ron Weasley", 
-           "Fred Weasley", 
-           "George Weasley", 
-           "Percy Weasley", 
-           "Charlie Weasley", 
-           "Bill Weasley", 
-           "Hermione Granger", 
-           "Fleur Delacour", 
-           "Gabrielle Delacour", 
-           "Audrey UNKNOWN", 
-           "James Potter II", 
-           "Albus Potter", 
-           "Lily Potter", 
-           "Rose Weasley", 
-           "Hugo Weasley", 
-           "Victoire Weasley", 
-           "Dominique Weasley", 
-           "Louis Weasley", 
-           "Molly Weasley", 
+  name = c("Vernon Dursley",
+           "Marjorie Dursley",
+           "Petunia Evans",
+           "Lily Evans",
+           "James Potter",
+           "Dudley Dursley",
+           "Harry Potter",
+           "Ginny Weasley",
+           "Arthur Weasley",
+           "Molly Prewett",
+           "Ron Weasley",
+           "Fred Weasley",
+           "George Weasley",
+           "Percy Weasley",
+           "Charlie Weasley",
+           "Bill Weasley",
+           "Hermione Granger",
+           "Fleur Delacour",
+           "Gabrielle Delacour",
+           "Audrey UNKNOWN",
+           "James Potter II",
+           "Albus Potter",
+           "Lily Potter",
+           "Rose Weasley",
+           "Hugo Weasley",
+           "Victoire Weasley",
+           "Dominique Weasley",
+           "Louis Weasley",
+           "Molly Weasley",
            "Lucy Weasley"),
   gen = c(1, 1, 1, 1, 1, 2, 2, 2, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3),
[…]
