+++ data-raw/test_hp_ped_network.Rmd
@@ -1,6 +1,7 @@
 ---
-title: "Network tools for finding extended pedigrees and path tracing"
+title: "Harry Potter Pedigree"
 output: rmarkdown::html_vignette
+author: "Rachel Good"
 vignette: >
   %\VignetteIndexEntry{Network}
   %\VignetteEngine{knitr::rmarkdown}
@@ -12,6 +13,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+
+
 options(rmarkdown.html_vignette.check_title = FALSE)
 ```
 
@@ -41,13 +44,6 @@ data(hp_ped)
 
 # Finding Extended Families
 
-Many pedigree datasets only contain information on the person, their
-mother, and their father, often without nuclear or extended family IDs.
-Recognizing which sets of people are unrelated simplifies many
-pedigree-related tasks. This function facilitates those tasks by finding
-all the extended families. People within the same extended family have
-at least some form of relation, however distant, while those in
-different extended families have no relations.
 
 ```{r, echo=FALSE, results='hide', out.width='50%', fig.cap="hp_ped Pedigree"}
 plotPedigree(hp_ped, code_male = 0, verbose = TRUE, )
@@ -146,4 +142,24 @@ the `hp_ped` dataset through sharing an extended family.
 extendedFamilyEnvironment <- ped2ce(hp_ped)
 extendedFamilyEnvironment[1:7, 1:7]
 table(extendedFamilyEnvironment)
+```
+
+
+
+# Tests
+```{r}
+plotPedigree(hp_ped, width = 1,code_male = "M",verbose = FALSE)
+
+```
+
+```{r}
+data(inbreeding)
+df <- inbreeding
+summary(df)
+plotPedigree(df, width = 1)
+
+data(hazard)
+df <- hazard
+summary(df)
+plotPedigree(df, width = 1)
 ```
