### 2023-11-02 85dfc0c
Update validatePedigree.R
