+++ R/validatePedigree.R
@@ -16,8 +16,8 @@ standardize_colnames <- function(df, verbose = FALSE) {
     "fam" = "^(?:fam(?:ily)?(?:id)?)",
     "ID" = "^(?:i(?:d$|ndiv(?:idual)?)|p(?:erson)?id)",
     "gen" = "^(?:gen(?:s|eration)?)",
-    "dadID" = "^(?:d(?:ad)?id|fatherid)",
-    "momID" = "^(?:m(?:om)?id|motherid)",
+    "dadID" = "^(?:d(?:ad)?id|paid|fatherid)",
+    "momID" = "^(?:m(?:om|a)?id|motherid)",
     "spt" = "^(?:s(?:pt)?id|spouse(?:id)?)",
     "twinID" = "^(?:twin(?:id)?)",
     "sex" = "^(?:sex|gender|female|m(?:a(?:le|n)|en)|wom[ae]n)"
