+++ data-raw/HP_df.R
@@ -6,10 +6,10 @@ library(tidyverse)
 ## Create dataframe
 
 hp_ped <- data.frame(
-  ID = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
+  personID = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
          11, 12, 13, 14, 15, 16, 17, 18, 19, 20,
          21, 22, 23, 24, 25, 26, 27, 28, 29, 30),
-  FamID = c("fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1", "fam1"),
+  fam = c("1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1"),
   name = c("Vernon Dursley",
            "Marjorie Dursley",
            "Petunia Evans",
@@ -41,11 +41,11 @@ hp_ped <- data.frame(
            "Molly Weasley",
            "Lucy Weasley"),
   gen = c(1, 1, 1, 1, 1, 2, 2, 2, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3),
-  motherID = c(101, 101, 103, 103, NA, 3, 4, 10, NA, NA, 10, 10, 10, 10, 10, 10, NA, 105, 105, NA, 8, 8, 8, 17, 17, 18, 18, 18, 20, 20),
-  fatherID = c(102, 102, 104, 104, NA, 1, 5, 9, NA, NA, 9, 9, 9, 9, 9, 9, NA, 106, 106, NA, 7, 7, 7, 11, 11, 16, 16, 16, 14, 14),
+  momID = c(101, 101, 103, 103, NA, 3, 4, 10, NA, NA, 10, 10, 10, 10, 10, 10, NA, 105, 105, NA, 8, 8, 8, 17, 17, 18, 18, 18, 20, 20),
+  dadID = c(102, 102, 104, 104, NA, 1, 5, 9, NA, NA, 9, 9, 9, 9, 9, 9, NA, 106, 106, NA, 7, 7, 7, 11, 11, 16, 16, 16, 14, 14),
   spouseID = c(3, NA, 1, 5, 4, NA, 8, 7, 10, 9, 17, NA, NA, 20, NA, 18, 11, 16, NA, 14, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA),
   sex = c(1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 0)
-) #sex: 0 = female, 1 = male 
+) #sex: 0 = female, 1 = male
 
 
 # audrey weasley - unknown maiden name
@@ -56,7 +56,7 @@ hp_ped <- data.frame(
 
 
 hp_ped[nrow(hp_ped) + 1,] <- list(101,
-                                  "fam1",
+                                  "1",
                                   "Mother Durseley",
                                   0,
                                   NA,
@@ -64,7 +64,7 @@ hp_ped[nrow(hp_ped) + 1,] <- list(101,
                                   102,
                                   0)
 hp_ped[nrow(hp_ped) + 1,] <- list(102,
-                                  "fam1",
+                                  "1",
[…]
