### 2023-11-03 69e2b11
write csv
trying to get the data in the right place
