+++ man/standardize_colnames.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/validatePedigree.R
+% Please edit documentation in R/cleanPedigree.R
 \name{standardize_colnames}
 \alias{standardize_colnames}
 \title{Standardize Column Names in a Dataframe (Internal)}
