+++ NAMESPACE
@@ -10,6 +10,7 @@ export(famSizeCal)
 export(fitComponentModel)
 export(identifyComponentModel)
 export(inferRelatedness)
+export(makeInbreeding)
 export(makeTwins)
 export(ped2add)
 export(ped2ce)
