+++ R/tweakPedigree.R
@@ -16,13 +16,15 @@ makeTwins <- function(ped, ID_twin1 = NA_integer_, ID_twin2 = NA_integer_, gen_t
   # a support function
   resample <- function(x, ...) x[sample.int(length(x), ...)]
   # Check if the ped is the same format as the output of simulatePedigree
-  if (paste0(colnames(ped), collapse = "") != paste0(c("fam", "ID", "gen",
-   "dadID", "momID", "spt", "sex"), collapse = "")) {
+  if (paste0(colnames(ped), collapse = "") != paste0(c(
+    "fam", "ID", "gen",
+    "dadID", "momID", "spt", "sex"
+  ), collapse = "")) {
     ped <- standardize_colnames(ped)
-    if(verbose){
+    if (verbose) {
       cat("The input pedigree is not in the same format as the output of simulatePedigree\n")
     }
-    #stop("The input pedigree is not in the same format as the output of simulatePedigree")
+    # stop("The input pedigree is not in the same format as the output of simulatePedigree")
   }
   ped$MZtwin <- NA_integer_
   # Check if the two IDs are provided
@@ -88,7 +90,7 @@ makeTwins <- function(ped, ID_twin1 = NA_integer_, ID_twin2 = NA_integer_, gen_t
     ped$MZtwin[ped$ID == ID_twin1] <- ID_twin2
     ped$MZtwin[ped$ID == ID_twin2] <- ID_twin1
   }
-  if(verbose){
+  if (verbose) {
     cat("twin1", ID_twin1, "\n")
     cat("twin2", ID_twin2, "\n")
   }
@@ -98,7 +100,7 @@ makeTwins <- function(ped, ID_twin1 = NA_integer_, ID_twin2 = NA_integer_, gen_t
 
 #' makeInbreeding
 #' A function to create inbred mates in the simulated pedigree \code{data.frame}.
-#' Inbred mates can be created by specifying their IDs or by specifying the generation the inbred mate should be created. 
+#' Inbred mates can be created by specifying their IDs or by specifying the generation the inbred mate should be created.
 #' When specifying the generation, inbreeding between siblings or 1st cousin needs to be specified.
 #' This is a supplementary function for \code{simulatePedigree}.
 #' @param ped A \code{data.frame} in the same format as the output of \code{simulatePedigree}.
@@ -112,22 +114,23 @@ makeTwins <- function(ped, ID_twin1 = NA_integer_, ID_twin2 = NA_integer_, gen_t
 
 # A function to create inbred mates in the simulated pedigree.
 
-makeInbreeding = function(ped, 
-                          ID_mate1 = NA_integer_, 
-                          ID_mate2 = NA_interger_, 
-                          verbose = FALSE, 
-                          gen_inbred = 2, 
-                          type_inbred = "sib"){
-    # A support function
-    resample <- function(x, ...) x[sample.int(length(x), ...)]                        
-    # check if the ped is the same format as the output of simulatePedigree
-    if (paste0(colnames(ped), collapse = "") != paste0(c("fam", "ID", "gen", "dadID", "momID", "spt", "sex"), collapse = "")) {
+makeInbreeding <- function(ped,
+                           ID_mate1 = NA_integer_,
+                           ID_mate2 = NA_integer_,
+                           verbose = FALSE,
+                           gen_inbred = 2,
+                           type_inbred = "sib") {
+  # A support function
+  resample <- function(x, ...) x[sample.int(length(x), ...)]
[…]
