+++ R/convertPedigree.R
@@ -29,7 +29,7 @@ ped2com <- function(ped, component,
     )
   )
   # standardize colnames
-  standardize_colnames(ped) <- ped
+#  standardize_colnames(ped) <- ped
   # Get the number of rows in the pedigree dataset, representing the size of the family
   nr <- nrow(ped)
 
