### 2023-11-06 8174a6e
Revert "Merge branch 'dev' of https://github.com/R-Computing-Lab/BGmisc into dev"
This reverts commit 4b51c73884cbce67549c9085cc15da8b5c1ef8ba, reversing
changes made to 5d1ea2d38766ae421cb19ff7eec1d888f3a47bf9.
