### 2023-11-06 0aeab84
Revert "Merge branch 'dev' of https://github.com/R-Computing-Lab/BGmisc into dev"
This reverts commit 1549e457401df524420bbb46097177a1c2ecfd1f, reversing
changes made to 4b51c73884cbce67549c9085cc15da8b5c1ef8ba.
