+++ man/ped2com.Rd
@@ -12,6 +12,7 @@ ped2com(
   verbose = FALSE,
   gc = FALSE,
   flatten.diag = FALSE,
+  standardize_colnames = TRUE,
   ...
 )
 }
@@ -31,6 +32,8 @@ generations as there are in the data.}
 \item{gc}{logical. If TRUE, do frequent garbage collection via \code{\link{gc}} to save memory}
 
 \item{flatten.diag}{logical. If TRUE, overwrite the diagonal of the final relatedness matrix with ones}
+
+\item{standardize_colnames}{logical. If TRUE, standardize the column names of the pedigree dataset}
 
 \item{...}{additional arguments to be passed to \code{\link{ped2com}}}
 }
