### 2023-11-07 749cdf5
Update R-CMD-check.yaml
temp removal... 

          - {os: ubuntu-latest, r: 'devel', rtools: '', http-user-agent: 'release'}
          - {os: ubuntu-latest, r: 'release', rtools: ''}
          - {os: ubuntu-latest, r: 'oldrel-1', rtools: ''}
