+++ .github/workflows/R-CMD-check.yaml
@@ -22,9 +22,7 @@ jobs:
         config:
           - {os: macOS-latest, r: 'release', rtools: ''}
           - {os: windows-latest, r: 'release', rtools: '42'}
-          - {os: ubuntu-latest, r: 'devel', rtools: '', http-user-agent: 'release'}
-          - {os: ubuntu-latest, r: 'release', rtools: ''}
-          - {os: ubuntu-latest, r: 'oldrel-1', rtools: ''}
+
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
