### 2023-11-07 4542d7e
Revert "Update R-CMD-check.yaml"
This reverts commit 749cdf534a7cbf04d1860b87350e3eeafe014088.
