+++ man/repairIDs.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/checkIDs.R
+\name{repairIDs}
+\alias{repairIDs}
+\title{Repair Missing IDs}
+\usage{
+repairIDs(ped, verbose = FALSE)
+}
+\arguments{
+\item{ped}{A pedigree object}
+
+\item{verbose}{A logical indicating whether to print progress messages}
+}
+\value{
+A corrected pedigree
+}
+\description{
+This function repairs missing IDs in a pedigree.
+}
