### 2023-11-08 dd1cd16
Update man pages that should be .gitignore-d
