+++ man/inbreeding.Rd
@@ -18,7 +18,7 @@ Different kinds of inbreeding occur in each extended family.
 The types of inbreeding are as follows:
 
 \itemize{
-    \item{Extended Family 1}{Sister wives: Children with the same father and different mothers who are sisters}
+    \item{Extended Family 1. }{Sister wives: Children with the same father and different mothers who are sisters}
     \item{Extended Family 2. }{Full siblings have children}
     \item{Extended Family 3. }{Half siblings have children}
     \item{Extended Family 4. }{First cousins have children}
