### 2023-11-08 3d1f819
Add skeleton maternal/paternal line functions
