+++ R/family.R
@@ -21,8 +21,15 @@
 #' @export
 #'
 ped2fam <- function(ped, personID = "ID", momID = "momID", dadID = "dadID", famID = "famID") {
+  # Call to wrapper function
+  .ped2id(ped = ped, personID = personID, momID = momID, dadID = dadID, famID = famID, type = 'parents')
+}
+
+.ped2id <- function(ped,
+                    personID = "ID", momID = "momID", dadID = "dadID",
+                    famID = "famID", type) {
   # Turn pedigree into family
-  pg <- ped2graph(ped = ped, personID = personID, momID = momID, dadID = dadID, famID = famID)
+  pg <- ped2graph(ped = ped, personID = personID, momID = momID, dadID = dadID, adjacent = type)
 
   # Find weakly connected components of graph
   wcc <- igraph::components(pg)
@@ -33,20 +40,28 @@ ped2fam <- function(ped, personID = "ID", momID = "momID", dadID = "dadID", famI
   )
   names(fam) <- c(personID, famID)
   ped2 <- merge(fam, ped,
-    by = personID, all.x = FALSE, all.y = TRUE
+                by = personID, all.x = FALSE, all.y = TRUE
   )
 
   return(ped2)
 }
 
+
 #' Turn a pedigree into a graph
 #' @param ped a pedigree dataset.  Needs ID, momID, and dadID columns
 #' @inheritParams ped2fam
 #' @param directed Logical scalar. Default is TRUE. Indicates whether or not to create a directed graph.
+#' @param adjacent Character.  Relationship that defines adjacency in the graph: parents, mothers, or fathers
 #' @details
 #' The general idea of this function is to represent a pedigree as a graph using the igraph package.
 #'
 #' Once in graph form, several common pedigree tasks become much simpler.
+#'
+#' The \code{adjacent} argument allows for different kinds of graph structures.
+#' When using \code{parents} for adjacency, the graph shows all parent-child relationships.
+#' When using \code{mother} for adjacency, the graph only shows mother-child relationships.
+#' Similarly when using \code{father} for adjacency, only father-child relationships appear in the graph.
+#' Extended families can be constructed from the parent graph, maternal lines from the mothers graph, and paternal lines from the fathers graph.
 #'
 #' @returns
 #' A graph
@@ -57,20 +72,60 @@ ped2graph <- function(ped,
                       personID = "ID",
                       momID = "momID",
                       dadID = "dadID",
-                      famID = "famID",
-                      directed = TRUE) {
+                      directed = TRUE,
+                      adjacent = c('parents', 'mothers', 'fathers')) {
+  # Check ped/data.fram
   if (!inherits(ped, "data.frame")) stop("ped should be a data.frame or inherit to a data.frame")
-  if (!all(c() %in% names(ped))) stop("'personID', 'momID', and 'dadID' were not all found in your pedigree.\nMake sure you have the variable names correct.")
+  # Handle adjacent argument
+  adjacent <- match.arg(tolower(adjacent)[1],
+                         choices = c(
+                           "parents",
+                           "mothers",
+                           "fathers"
+                         )
+  )
+  # Check the needed IDs are in the data
+  if(adjacent == 'parents'){
+    needIds <- c(personID, momID, dadID)
+  } else if(adjacent == 'mothers'){
+    needIds <- c(personID, momID)
+  } else if(adjacent == 'fathers') {
+    needIds <- c(personID, dadID)
+  }
+  if (!all(c(needIds) %in% names(ped))){
[…]
