+++ NAMESPACE
@@ -19,7 +19,9 @@ export(ped2cn)
 export(ped2com)
 export(ped2fam)
 export(ped2graph)
+export(ped2maternal)
 export(ped2mit)
+export(ped2paternal)
 export(plotPedigree)
 export(related_coef)
 export(relatedness)
