+++ R/checkSex.R
@@ -5,20 +5,20 @@
 #' 2. Optionally repairs the sex coding based on a specified logic.
 #'
 #' @param ped A dataframe representing the pedigree data with a 'sex' column.
-#' @param code_male The code used to represent males in the 'sex' column. If NULL, no recoding is performed.
+#' @param code_male The code used to represent males in the 'sex' column. If both male and female are NULL, no recoding is performed.
+#' @param code_female The code used to represent females in the 'sex' column. If both male and female are NULL, no recoding is performed.
 #' @param verbose A logical flag indicating whether to print progress and validation messages to the console.
 #' @param repair A logical flag indicating whether to attempt repairs on the sex coding.
-#' @param recode A logical flag indicating whether to recode the 'sex' variable.
 #'
 #' @return Depending on the value of `repair`, either a list containing validation results or a repaired dataframe is returned.
 #' @examples
 #' \dontrun{
 #' ped <- data.frame(ID = c(1, 2, 3), sex = c("M", "F", "M"))
-#' checkSex(ped, code_male = "M", verbose = TRUE, repair = FALSE, recode = TRUE)
+#' checkSex(ped, code_male = "M", verbose = TRUE, repair = FALSE)
 #' }
 #' @export
 #'
-checkSex <- function(ped, code_male = NULL, verbose = FALSE, repair = FALSE, recode = FALSE) {
+checkSex <- function(ped, code_male = NULL, code_female = NULL,verbose = FALSE, repair = FALSE) {
   # Standardize column names in the input dataframe
   ped <- standardize_colnames(ped)
 
@@ -26,6 +26,11 @@ checkSex <- function(ped, code_male = NULL, verbose = FALSE, repair = FALSE, rec
 
   # Initialize a list to store validation results
   validation_results <- list()
+
+  # Initialize a list to track changes made during repair
+  changes <- list()
+
+
   if (verbose) {
     cat("Step 1: Checking how many genders...\n")
   }
@@ -35,23 +40,28 @@ checkSex <- function(ped, code_male = NULL, verbose = FALSE, repair = FALSE, rec
   validation_results$sex_length <- length(unique(ped$sex))
   if (verbose) {
     cat(paste0(
-      validation_results$sex_length, " unique values found.\n ",
-      paste0(validation_results$sex_unique)
-    ))
+      validation_results$sex_length, " unique values found.\n "
+    ),
+    paste0(validation_results$sex_unique))
   }
-
 
 
   if (repair) {
     if (verbose) {
-      cat("Step 3: Attempting to repair sex coding...\n")
+      cat("Step 2: Attempting to repair sex coding...\n")
     }
-    # Initialize a list to track changes made during repair
-    changes <- list()
+
     # [Insert logic to repair sex coding here]
     if (validation_results$sex_length == 2) {
       most_frequent_sex_dad <- names(sort(table(ped$sex[ped$ID %in% ped$dadID]), decreasing = TRUE))[1]
+      original_ped <- ped
       ped <- recodeSex(ped, code_male = most_frequent_sex_dad)
+      # Count and record the change
+      num_changes <- sum(original_ped$sex != ped$sex)
+      # Record the change and the count
+      changes[[length(changes) + 1]] <- sprintf(
+        "Recode sex based on most frequent sex in dads: %s. Total gender changes made: %d",
+        most_frequent_sex_dad, num_changes)
     }
     # Update the pedigree dataframe after repair
     repaired_ped <- ped
[…]
