### 2023-11-16 a9c48c1
Update checkSex.R
