+++ vignettes/network.Rmd
@@ -36,7 +36,7 @@ structure inherent in pedigrees:
 
 ```{r setup}
 library(BGmisc)
-data(hazard)
+data(potter)
 ```
 
 # Finding Extended Families
@@ -49,18 +49,19 @@ all the extended families. People within the same extended family have
 at least some form of relation, however distant, while those in
 different extended families have no relations.
 
-```{r, echo=FALSE, results='hide', out.width='50%', fig.cap="Hazard Pedigree"}
-plotPedigree(hazard, code_male = 0, verbose = TRUE, )
+```{r, echo=FALSE, results='hide', out.width='50%', fig.cap="potter Pedigree"}
+plotPedigree(potter, code_male = 1, verbose = TRUE)
 ```
 
-We will use the `hazard` pedigree data as an example.
+We will use the `potter` pedigree data as an example.
 
 ```{r}
-ds <- ped2fam(hazard, famID = "newFamID")
-table(ds$FamID, ds$newFamID)
+ds <- ped2fam(potter, famID = "newFamID", personID = "personID")
+table(ds$famID, ds$newFamID)
+
 ```
 
-Because the `hazard` data already had a family ID variable we compare
+Because the `potter` data already had a family ID variable we compare
 our newly created variable to the pre-existing one. They match!
 
 # Computing Relatedness
@@ -70,7 +71,7 @@ you'll likely want to know how much. For additive genetic relatedness,
 you can use the `ped2add()` function.
 
 ```{r}
-add <- ped2add(hazard)
+add <- ped2add(potter)
 ```
 
 This computes the additive genetic relatedness for everyone in the data.
@@ -96,9 +97,9 @@ family.
 
 ```{r}
 add_list <- lapply(
-  unique(hazard$FamID),
+  unique(potter$famID),
   function(d) {
-    tmp <- hazard[hazard$FamID %in% d, ]
+    tmp <- potter[potter$famID %in% d, ]
     ped2add(tmp)
   }
 )
@@ -113,10 +114,10 @@ extended family environment (`ped2ce`).
 ### Computing mitochondrial relatedness
 
 Here we calculate the mitochondrial relatedness between all pairs of
-individuals in the `hazard` dataset.
+individuals in the `potter` dataset.
 
 ```{r}
-mit <- ped2mit(hazard)
+mit <- ped2mit(potter)
 mit[1:7, 1:7]
 table(mit)
 ```
@@ -128,10 +129,10 @@ not.
 ### Computing relatedness through common nuclear environment
 
 Here we calculate the relatedness between all pairs of individuals in
-the `hazard` dataset through sharing both parents.
+the `potter` dataset through sharing both parents.
 
 ```{r}
-commonNuclear <- ped2cn(hazard)
+commonNuclear <- ped2cn(potter)
 commonNuclear[1:7, 1:7]
 
 table(commonNuclear)
@@ -140,10 +141,10 @@ table(commonNuclear)
 ### Computing relatedness through common extended family environment
 
 Here we calculate the relatedness between all pairs of individuals in
-the `hazard` dataset through sharing an extended family.
+the `potter` dataset through sharing an extended family.
 
 ```{r}
-extendedFamilyEnvironment <- ped2ce(hazard)
+extendedFamilyEnvironment <- ped2ce(potter)
 extendedFamilyEnvironment[1:7, 1:7]
 table(extendedFamilyEnvironment)
 ```
