+++ man/recodeSex.Rd
@@ -8,6 +8,7 @@ recodeSex(
   ped,
   verbose = FALSE,
   code_male = NULL,
+  code_female = NULL,
   code_na = NULL,
   recode_male = "M",
   recode_female = "F",
@@ -19,7 +20,9 @@ recodeSex(
 
 \item{verbose}{A logical flag indicating whether to print progress and validation messages to the console.}
 
-\item{code_male}{The code used to represent males in the 'sex' column. If NULL, no recoding is performed.}
+\item{code_male}{The code used to represent males in the 'sex' column. If both male and female are NULL, no recoding is performed.}
+
+\item{code_female}{The code used to represent females in the 'sex' column. If both male and female are NULL, no recoding is performed.}
 }
 \value{
 A modified version of the input data.frame \code{ped}, containing an additional or modified 'sex_recode' column where the 'sex' values are recoded according to \code{code_male}. NA values in the 'sex' column are preserved.
