### 2023-11-22 0027452
Update convertPedigree.R
added two t cross product solutions
