### 2023-11-22 903dfb4
Update convertPedigree.R
added two t cross product solutions
