+++ man/ped2mit.Rd
@@ -12,7 +12,9 @@ ped2mit(
   verbose = FALSE,
   gc = FALSE,
   flatten.diag = FALSE,
-  standardize.colnames = TRUE
+  standardize.colnames = TRUE,
+  alt.tcross1 = FALSE,
+  alt.tcross2 = FALSE
 )
 }
 \arguments{
@@ -31,6 +33,10 @@ generations as there are in the data.}
 \item{flatten.diag}{logical. If TRUE, overwrite the diagonal of the final relatedness matrix with ones}
 
 \item{standardize.colnames}{logical. If TRUE, standardize the column names of the pedigree dataset}
+
+\item{alt.tcross1}{logical. If TRUE, use an alternative method for computing the transpose}
+
+\item{alt.tcross2}{logical. If TRUE, use an alternative method for computing the transpose}
 }
 \description{
 Take a pedigree and turn it into a mitochondrial relatedness matrix
