+++ man/ped2com.Rd
@@ -13,6 +13,8 @@ ped2com(
   gc = FALSE,
   flatten.diag = FALSE,
   standardize.colnames = TRUE,
+  alt.tcross1 = FALSE,
+  alt.tcross2 = FALSE,
   ...
 )
 }
@@ -34,6 +36,10 @@ generations as there are in the data.}
 \item{flatten.diag}{logical. If TRUE, overwrite the diagonal of the final relatedness matrix with ones}
 
 \item{standardize.colnames}{logical. If TRUE, standardize the column names of the pedigree dataset}
+
+\item{alt.tcross1}{logical. If TRUE, use an alternative method for computing the transpose}
+
+\item{alt.tcross2}{logical. If TRUE, use an alternative method for computing the transpose}
 
 \item{...}{additional arguments to be passed to \code{\link{ped2com}}}
 }
