### 2023-11-22 4b40bf7
Revert "Update convertPedigree.R"
This reverts commit 00274521828abc3050c5de4f8ae38258d7d75510.
