### 2023-11-27 6b05d47
Work on 5-person plot
Changed var names; added n_rows to plotpedigree function
