### 2023-11-27 da2cdc5
5-person plot
Added 5-person subplot
