### 2023-11-27 a363336
Add option to randomly generate number of kids per mate
