+++ R/simulatePedigree.R
@@ -10,6 +10,7 @@
 #' @param Ngen Number of generations. An integer >= 2 that determines how many generations the simulated pedigree will have. The first generation is always a fertilized couple. The last generation has no mated individuals.
 #' @param sexR Sex ratio of offspring. A numeric value ranging from 0 to 1 that determines the proportion of males in all offspring in this pedigree. For instance, 0.4 means 40 percent of the offspring will be male.
 #' @param marR Mating rate. A numeric value ranging from 0 to 1 which determines the proportion of mated (fertilized) couples in the pedigree within each generation. For instance, marR = 0.5 suggests 50 percent of the offspring in a specific generation will be mated and have their offspring.
+#' @param rd_kpc logical. If TRUE, the number of kids per mate will be randomly generated from a poisson distribution with mean kpc. If FALSE, the number of kids per mate will be fixed at kpc.
 #' @param balancedSex Not fully developed yet. Always \code{TRUE} in the current version.
 #' @param balancedMar Not fully developed yet. Always \code{TRUE} in the current version.
 #' @param verbose logical  If TRUE, print progress through stages of algorithm
@@ -30,6 +31,7 @@ simulatePedigree <- function(kpc = 3,
                              Ngen = 4,
                              sexR = .5,
                              marR = 2 / 3,
+                             rd_kpc = FALSE,
                              balancedSex = TRUE,
                              balancedMar = TRUE,
                              verbose = FALSE) {
@@ -236,29 +238,17 @@ simulatePedigree <- function(kpc = 3,
       # get the number of linked female and male children after excluding the single children
       # get a vector of single person id in the ith generation
       IdSingle <- df_Ngen$id[is.na(df_Ngen$spt)]
-      # print(N_LinkedMem)
-      # print(N_LinkedFemale)
-      # print(N_LinkedMale)
       SingleF <- sum(df_Ngen$sex == "F" & is.na(df_Ngen$spt))
       CoupleF <- N_LinkedFemale - SingleF
-      # print(CoupleF)
-
       SingleM <- sum(df_Ngen$sex == "M" & is.na(df_Ngen$spt))
       CoupleM <- N_LinkedMale - SingleM
-      # print(CoupleM)
       # get all couple ids
-      # print(i)
-      # print(df_Ngen$IdCouple)
       coupleID <- unique(df_Ngen$IdCouple[!is.na(df_Ngen$IdCouple)])
-      # print(coupleID)
       if (i == Ngen) {
         CoupleF <- 0
       }
       coupleGirl <- sample(coupleID, CoupleF)
-      # print(coupleGirl)
       coupleBoy <- coupleID[!coupleID %in% coupleGirl]
-      # print(coupleGirl)
-      # print(coupleBoy)
       # single person should all be sons or daus
       # change the ifson and ifdau based on coupleGirl and coupleBoy
       for (j in 1:sizeGens[i]) {
@@ -280,8 +270,6 @@ simulatePedigree <- function(kpc = 3,
           }
         }
       }
-      # print(df_Ngen)
-
 
       df_Ngen <- df_Ngen[order(as.numeric(rownames(df_Ngen))), , drop = FALSE]
       df_Ngen <- df_Ngen[, -ncol(df_Ngen)]
@@ -337,10 +325,34 @@ simulatePedigree <- function(kpc = 3,
         # print(IdDau)
         IdOfp <- evenInsert(IdSon, IdDau)
 
+        if (rd_kpc) {
+          # get the number of pairs of mates in the i th generation
+          nMates <- sum(df_Ngen$ifparent)/2
[…]
