### 2023-12-18 d61daa6
Update potter vignette for new variable names
