### 2023-12-19 5fde91a
Add ... to allow passing further named args to kinship2 function
