### 2024-01-10 #5894 @mhu48
Happy to do so, look like an interesting contribution!

Øystein Sørensen ***@***.***> 于 2024年1月7日周日 14:05写道：

> 👋 @mhu48 <https://github.com/mhu48>, @vankesteren
> <https://github.com/vankesteren>, @wjakethompson
> <https://github.com/wjakethompson>, would any of you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
