### 2024-01-10 #5894 @osorensen
> I don't have any experience with behavior genetics, so I'm would not be able to give feedback on the accuracy/quality of the methods. I could review for R package best practices (e.g., test coverage, documentation, etc.) if that would be helpful.

Thanks @wjakethompson, that would be really helpful.
