### 2024-01-12 1c4aa26
Random Pedigree Done and further debugged. Add dropLink funtion
