### 2024-02-02 f68d875
Add CRAN installation instructions to README
