### 2024-02-02 R-Computing-Lab/BGmisc#12 @wjakethompson
Increase test coverage
JOSS review criterion: Automated tests (openjournals/joss-reviews#6203)

Using `covr::package_coverage()`, I see 62.94% test coverage. This is a little lower than I think is preferred. The output below indicates the coverage by file, and there are several files with 0 or very low coverage. Not being familiar the subject matter of the package, I'm not sure whether it would be possible to write additional tests for these functions, but is something to consider.

I'd also recommend automating test coverage with a GitHub action. Right now, tests will be run as part of the R CMD CHECK action, which would indicate if a test was newly failing. However, adding a test coverage action would allow you to monitor how new PRs affect overall test coverage, which could then also be reported with a badge in the README. `usethis::use_github_action("test-coverage")` will set this up.

```r
covr::package_coverage()

#> BGmisc Coverage: 62.94%
#> R/checkIDs.R: 0.00%
#> R/tweakPedigree.R: 0.00%
#> R/checkSex.R: 12.12%
#> R/helper.R: 28.57%
#> R/identification.R: 48.15%
#> R/famSizeCal.R: 50.00%
#> R/formula.R: 57.14%
#> R/family.R: 60.00%
#> R/evenInsert.R: 66.67%
#> R/plotPedigree.R: 83.33%
#> R/convertPedigree.R: 88.15%
#> R/simulatePedigree.R: 93.40%
#> R/cleanPedigree.R: 95.00%
```
