### 2024-02-02 R-Computing-Lab/BGmisc#13 @wjakethompson
Code functionality issues
JOSS review criterion: Functionality (openjournals/joss-reviews#6203)

`goodpractice` pick up several issues that may impact package functionality.

1. Test coverage (already addressed in #12)
2. `simulatePedigree()` has a very high cyclomatic complexity. In general, high complexity indicates that code is less legible, more difficult to evaluate for the correctness of the implementation, and therefore more difficult to maintain. Maybe try refactoring or modularizing this function to get complexity down to more reasonable values (e.g., 20-50)? I'm not familiar with the subject matter here, so this may or may not actually be actionable.
3. `sapply()` is not type safe and may provide unexpected results in edge cases.
4. Using `seq_along()` and `seq_length()` also avoids errors in edge cases that are introduced by, for example, 1:nrow(...).

There are also a few other notes from `goodpractice` included below tha twould be helpful, but are less critical.

```
goodpractice::gp()

── GP BGmisc ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 62% of code lines are covered by
    test cases.

    R/checkIDs.R:20:NA
    R/checkIDs.R:23:NA
    R/checkIDs.R:25:NA
    R/checkIDs.R:26:NA
    R/checkIDs.R:30:NA
    ... and 287 more lines

  ✖ write short and simple functions. These functions have high cyclomatic complexity (>50):
    simulatePedigree (72). You can make them easier to reason about by encapsulating distinct steps of your
    function into subfunctions.
  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite often. A build date will be
    added to the package when you perform `R CMD build` on it.
  ✖ add a "URL" field to DESCRIPTION. It helps users find information about your package online. If your
    package does not have a homepage, add an URL to GitHub, or the CRAN package package page.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug tracker. Many online code hosting
    services provide bug trackers for free, https://github.com, https://gitlab.com, etc.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are
    about 80 characters wide. Try make your lines shorter than 80 characters

    data-raw/HP_df.R:15:81
    data-raw/HP_df.R:48:81
    data-raw/HP_df.R:49:81
    data-raw/HP_df.R:50:81
    data-raw/HP_df.R:51:81
    ... and 228 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data.
    Consider using vapply() instead.

    R/family.R:111:17
    R/family.R:112:17
    R/family.R:120:24
    R/family.R:127:24
    R/helper.R:31:7

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions. They are error
    prone and result 1:0 if the expression on the right hand side is zero. Use seq_len() or seq_along() instead.

    R/evenInsert.R:23:13
    R/evenInsert.R:29:15
    R/identification.R:28:27
    R/plotPedigree.R:79:25
    R/simulatePedigree.R:130:17
    ... and 4 more lines

  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead,
    import only the specific functions you need.
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```
