### 2024-02-02 #6203 @wjakethompson
I've complete my review. Just a few notes:

* I think installation instruction should include CRAN in addition to GitHub. I opened a PR with a small tweak to the README. crossref: R-Computing-Lab/BGmisc#11
* Test coverage is a little low (~63%) and could be better automated. crossref: R-Computing-Lab/BGmisc#12
* A few recommendations from `goodpractice::gp()` to ensure functionality of the package works as intended. crossref: R-Computing-Lab/BGmisc#13
