### 2024-02-02 R-Computing-Lab/BGmisc#13 @smasongarrison
Thanks! These are great suggestions. I have a suspicion that modularizing the pedigree simulation might not be feasible, but I'll ping the lead on that function @Xuanyu-Lyu For the rest of it, I'm happy to implement them :)
