### 2024-02-02 9c24796
Update NEWS.md
spelling
