+++ NEWS.md
@@ -1,5 +1,5 @@
 # BGmisc 1.1.0
-* Added ability to simulate twins, trance paternal and maternal lines, and harry potter pedigree
+* Added ability to simulate twins, trace paternal and maternal lines, and Harry Potter pedigree
 
 # BGmisc 1.0.1
 * Hot fix to resolve plotPedigree wrapper function breaking for pedigrees that contained multiple families
