### 2024-02-03 #6203 @mhu48
I have completed my review. I think overall this is very well-done, but would point out some things to improve.
1. The software paper needs some improvement in order to make it more accessible to readers who may be not as familiar with the field.
2. The instructions for installation need to include CRAN, and be made more specific to R.
3. I second the opinion from @wjakethompson about the automated tests.
