### 2024-02-03 #6203 @smasongarrison
>The software paper needs some improvement in order to make it more accessible to readers who may be not as familiar with the field.

Thanks! Could you be a little more specific about what information/improvements you'd want to see, as well as clarify which "field" you're referring to? (I just want to make sure that I'm understanding where the gaps are)
