### 2024-02-05 4f3519b
Update potter vignette for new variable names
