+++ vignettes/network.Rmd
@@ -36,7 +36,7 @@ structure inherent in pedigrees:
 
 ```{r setup}
 library(BGmisc)
-data(hazard)
+data(potter)
 ```
 
 # Finding Extended Families
@@ -49,18 +49,24 @@ all the extended families. People within the same extended family have
 at least some form of relation, however distant, while those in
 different extended families have no relations.
 
-```{r, echo=FALSE, results='hide', out.width='50%', fig.cap="Hazard Pedigree"}
-plotPedigree(hazard, code_male = 0, verbose = TRUE, )
+```{r, echo=FALSE, results='hide', out.width='50%', fig.cap="Potter Family Pedigree"}
+plotPedigree(potter, code_male = 1, verbose = TRUE)
 ```
 
-We will use the `hazard` pedigree data as an example.
+We will use the `potter` pedigree data as an example. For convienence, we've renamed the family ID variable to `oldfam` to avoid confusion with the new family ID variable we will create.
 
 ```{r}
-ds <- ped2fam(hazard, famID = "newFamID")
-table(ds$FamID, ds$newFamID)
+
+names(potter)[names(potter)=="famID"]<- "oldfam"
+
+ds <- ped2fam(potter, famID = "oldfam", personID = "personID")
+
+
+table(ds$famID, ds$fam)
+
 ```
 
-Because the `hazard` data already had a family ID variable we compare
+Because the `potter` data already had a family ID variable, we compare
 our newly created variable to the pre-existing one. They match!
 
 # Computing Relatedness
@@ -70,7 +76,7 @@ you'll likely want to know how much. For additive genetic relatedness,
 you can use the `ped2add()` function.
 
 ```{r}
-add <- ped2add(hazard)
+add <- ped2add(potter)
 ```
 
 This computes the additive genetic relatedness for everyone in the data.
@@ -83,7 +89,7 @@ add[1:7, 1:7]
 
 The entry in the ith row and the jth column gives the relatedness
 between person i and person j. For example, person 1 and person 11 share
-`r add[1,11]`
+`r add[1,11]` of their nuclear DNA.
 
 ```{r}
 table(add)
@@ -96,9 +102,9 @@ family.
 
 ```{r}
 add_list <- lapply(
-  unique(hazard$FamID),
+  unique(potter$famID),
   function(d) {
-    tmp <- hazard[hazard$FamID %in% d, ]
+    tmp <- potter[potter$famID %in% d, ]
     ped2add(tmp)
   }
 )
@@ -112,11 +118,10 @@ extended family environment (`ped2ce`).
 
 ### Computing mitochondrial relatedness
 
-Here we calculate the mitochondrial relatedness between all pairs of
-individuals in the `hazard` dataset.
+Here we calculate the mitochondrial relatedness between all pairs of individuals in the `potter` dataset.
 
 ```{r}
-mit <- ped2mit(hazard)
+mit <- ped2mit(potter)
 mit[1:7, 1:7]
 table(mit)
 ```
@@ -128,10 +133,10 @@ not.
 ### Computing relatedness through common nuclear environment
 
 Here we calculate the relatedness between all pairs of individuals in
-the `hazard` dataset through sharing both parents.
+the `potter` dataset through sharing both parents.
 
 ```{r}
-commonNuclear <- ped2cn(hazard)
+commonNuclear <- ped2cn(potter)
 commonNuclear[1:7, 1:7]
 
 table(commonNuclear)
@@ -140,10 +145,10 @@ table(commonNuclear)
 ### Computing relatedness through common extended family environment
 
 Here we calculate the relatedness between all pairs of individuals in
-the `hazard` dataset through sharing an extended family.
[…]
