### 2024-02-05 293f095
Update network.Rmd
Updated for potter dataset
