### 2024-02-05 d48573e
Update network.Rmd
found it. moved it to a section that makes more sense
