+++ vignettes/network.Rmd
@@ -155,3 +155,15 @@ extendedFamilyEnvironment <- ped2ce(potter)
 extendedFamilyEnvironment[1:7, 1:7]
 table(extendedFamilyEnvironment)
 ```
+
+# Subsetting pedigrees
+
+Subsets of pedigrees can also be plotted to zoom in on smaller portions of pedigrees.
+However, this subset does not plot the relationship between spouses (such as the marriage between Vernon (1) and Petunia (3)), as there are not children to connect the two individuals together yet
+
+```{r, echo=FALSE, results='hide', out.width='50%', fig.cap="Potter Pedigree"}
+subset_rows <- c(1:5, 31:36)
+subset_potter <- potter[subset_rows, ]
+
+plotPedigree(subset_potter, code_male = 1, verbose = TRUE )
+```
