+++ vignettes/network.Rmd
@@ -17,8 +17,7 @@ options(rmarkdown.html_vignette.check_title = FALSE)
 
 # Introduction
 
-This vignette showcases two key features that capitalize on the network
-structure inherent in pedigrees:
+This vignette showcases two key features that capitalize on the network structure inherent in pedigrees:
 
 1.  Finding extended families with *any* connecting relationships
     between members. This feature strictly uses a person's ID, mother's
@@ -56,7 +55,7 @@ different extended families have no relations.
 plotPedigree(potter, code_male = 1, verbose = TRUE)
 ```
 
-We will use the `potter` pedigree data as an example. For convienence, we've renamed the family ID variable to `oldfam` to avoid confusion with the new family ID variable we will create.
+We will use the `potter` pedigree data as an example. For convenience, we've renamed the family ID variable to `oldfam` to avoid confusion with the new family ID variable we will create.
 
 ```{r}
 names(potter)[names(potter)=="famID"]<- "oldfam"
@@ -156,14 +155,34 @@ extendedFamilyEnvironment[1:7, 1:7]
 table(extendedFamilyEnvironment)
 ```
 
-# Subsetting pedigrees
+# Subsetting Pedigrees
 
-Subsets of pedigrees can also be plotted to zoom in on smaller portions of pedigrees.
-However, this subset does not plot the relationship between spouses (such as the marriage between Vernon (1) and Petunia (3)), as there are not children to connect the two individuals together yet
+Subsetting a pedigree allows researchers to focus on specific family lines or individuals within a larger dataset. This can be particularly useful for data validation as well as simplifying complex pedigrees for visualization. However, subsetting a pedigree can result in the underestimation of relatedness between individuals. This is because the subsetted pedigree may not contain all the individuals that connect two people together. For example if we were to remove Arthur Weasley (person 9) and Molly Prewett (person 10) from the `potter` dataset, we would lose the connections amoungst their children. 
+
 
 ```{r, echo=FALSE, results='hide', out.width='50%', fig.cap="Potter Pedigree"}
+names(potter)[names(potter)=="oldfam"]<- "famID"
+subset_rows <- c(1:8,11:36)
+subset_potter <- potter[subset_rows, ]
+
+subset_potter$dadID[subset_potter$dadID%in%c(9,10)] <- NA
+subset_potter$momID[subset_potter$momID%in%c(9,10)] <- NA
+
+plotPedigree(subset_potter, code_male = 1, verbose = TRUE )
+
+```
+
+In the plot above, we have removed Arthur Weasley (person 9) and Molly Prewett (person 10) from the `potter` dataset. As a result, the connections between their children are lost. 
+
+Similarly, if we remove the children of Vernon Dursey (1) and Petunia Evans(3) from the `potter` dataset, we would lose the connections between the two individuals.
+
+However, this subset does not plot the relationship between spouses (such as the marriage between ), as there are not children to connect the two individuals together yet.
+
+```{r}
 subset_rows <- c(1:5, 31:36)
 subset_potter <- potter[subset_rows, ]
+```
+```{r, echo=FALSE, results='hide', out.width='50%', fig.cap="Potter Pedigree"}
 
 plotPedigree(subset_potter, code_male = 1, verbose = TRUE )
 ```
