+++ .github/workflows/R-CMD-dev_maincheck.yaml
@@ -1,0 +1,54 @@
+on:
+  push:
+    branches: [dev_main]
+    paths:
+      - '**.Rmd'
+      - '**.Rd'
+      - '**.R'
+  pull_request:
+    branches: [dev_main]
+
+name: Dev Main Branch
+
+jobs:
+  R-CMD-check:
+    runs-on: ${{ matrix.config.os }}
+
+    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
+
+    strategy:
+      fail-fast: false
+      matrix:
+        config:
+          - {os: macOS-latest, r: 'release', rtools: ''}
+          - {os: windows-latest, r: 'release', rtools: '42'}
+          - {os: ubuntu-latest, r: 'devel', rtools: '', http-user-agent: 'release'}
+          - {os: ubuntu-latest, r: 'release', rtools: ''}
+          - {os: ubuntu-latest, r: 'oldrel-1', rtools: ''}
+
+    env:
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      R_KEEP_PKG_SOURCE: yes
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - uses: r-lib/actions/setup-pandoc@v2
+
+      - uses: r-lib/actions/setup-r@v2
+        with:
+          r-version: ${{ matrix.config.r }}
+          rtools-version: ${{ matrix.config.rtools }}
+          http-user-agent: ${{ matrix.config.http-user-agent }}
+          use-public-rspm: true
+
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          cache-version: 3
+          extra-packages: |
+            any::rcmdcheck
+          upgrade: 'TRUE'
+          needs: check
+      - uses: r-lib/actions/check-r-package@v2
+        with:
+          upload-snapshots: true
