+++ DESCRIPTION
@@ -14,7 +14,7 @@ Description: The BGmisc R package offers a comprehensive suite of functions tail
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 Imports: 
     Matrix,
     stats,
