### 2024-02-05 36d8525
added easy tests
added checks that verbose works and that the math solutions work
