+++ tests/testthat/test-network.R
@@ -159,3 +159,9 @@ test_that("ped2ce produces correct matrix dims, values, and dimnames", {
   expect_equal(dn[[1]], dn[[2]])
   expect_equal(dn[[1]], as.character(inbreeding$ID))
 })
+
+test_that("ped2add verbose prints updates",{
+  data(hazard)
+  expect_output(ped2add(hazard, verbose = TRUE), regexp = "Family Size =")
+})
+
