### 2024-02-06 2f3ba64
implementing predictable solutions
implementing testing
