### 2024-02-06 cf3b8f4
Update HP_df.R
updated potter var names to support structure of plotpedigree function
