### 2024-02-06 1b01e3e
Update test_checkSex.R
