+++ R/checkSex.R
@@ -5,27 +5,29 @@
 #' 2. Optionally repairs the sex coding based on a specified logic.
 #'
 #' @param ped A dataframe representing the pedigree data with a 'sex' column.
-#' @param code_male The code used to represent males in the 'sex' column. If NULL, no recoding is performed.
+#' @param code_male The code used to represent males in the 'sex' column. If both male and female are NULL, no recoding is performed.
+#' @param code_female The code used to represent females in the 'sex' column. If both male and female are NULL, no recoding is performed.
 #' @param verbose A logical flag indicating whether to print progress and validation messages to the console.
 #' @param repair A logical flag indicating whether to attempt repairs on the sex coding.
-#' @param recode A logical flag indicating whether to recode the 'sex' variable.
 #'
 #' @return Depending on the value of `repair`, either a list containing validation results or a repaired dataframe is returned.
 #' @examples
 #' \dontrun{
 #' ped <- data.frame(ID = c(1, 2, 3), sex = c("M", "F", "M"))
-#' checkSex(ped, code_male = "M", verbose = TRUE, repair = FALSE, recode = TRUE)
+#' checkSex(ped, code_male = "M", verbose = TRUE, repair = FALSE)
 #' }
 #' @export
 #'
-checkSex <- function(ped, code_male = NULL, verbose = FALSE, repair = FALSE, recode = FALSE) {
+checkSex <- function(ped, code_male = NULL, code_female = NULL,verbose = FALSE, repair = FALSE) {
   # Standardize column names in the input dataframe
   ped <- standardize_colnames(ped)
 
-  # bypass the rest of the function if recode_only is TRUE
+  # TO DO bypass the rest of the function if recode_only is TRUE
 
   # Initialize a list to store validation results
   validation_results <- list()
+
+
   if (verbose) {
     cat("Step 1: Checking how many genders...\n")
   }
@@ -39,19 +41,50 @@ checkSex <- function(ped, code_male = NULL, verbose = FALSE, repair = FALSE, rec
       paste0(validation_results$sex_unique)
     ))
   }
+ # are there multiple genders in the list of dads and moms?
 
+
+  table_sex_dad <- sort(table(ped$sex[ped$ID %in% ped$dadID]), decreasing = TRUE)
+  table_sex_mom <- sort(table(ped$sex[ped$ID %in% ped$momID]), decreasing = TRUE)
+  validation_results$all_sex_dad <- names(table_sex_dad)
+  validation_results$all_sex_mom <- names(table_sex_mom)
+  validation_results$most_frequent_sex_dad <- validation_results$all_sex_dad[1]
+  validation_results$most_frequent_sex_mom <- validation_results$all_sex_mom[1]
+
+# list ids for dads that are female, moms that are male
+if(length(validation_results$all_sex_dad) > 1){
+  df_dads <- ped[ped$ID %in% ped$dadID,]
+  validation_results$ID_female_dads <- df_dads$ID[df_dads$sex!=validation_results$most_frequent_sex_dad]
+  validation_results$ID_child_female_dads <- ped$ID[ped$dadID %in% validation_results$ID_female_dads]
+  remove(df_dads)
+}
+  if(length(validation_results$all_sex_mom) > 1){
+    df_moms <- ped[ped$ID %in% ped$momID,]
+    validation_results$ID_male_moms <- df_moms$ID[df_moms$sex!=validation_results$most_frequent_sex_mom]
+    validation_results$ID_child_male_moms <- ped$ID[ped$momID %in% validation_results$ID_female_moms]
+    remove(df_moms)
+  }
 
 
   if (repair) {
     if (verbose) {
-      cat("Step 3: Attempting to repair sex coding...\n")
[…]
