### 2024-02-09 3dd1b82
Update checkSex.R
