+++ man/recodeSex.Rd
@@ -9,6 +9,7 @@ recodeSex(
   verbose = FALSE,
   code_male = NULL,
   code_na = NULL,
+  code_female = NULL,
   recode_male = "M",
   recode_female = "F",
   recode_na = NA_character_
@@ -20,6 +21,8 @@ recodeSex(
 \item{verbose}{A logical flag indicating whether to print progress and validation messages to the console.}
 
 \item{code_male}{The code used to represent males in the 'sex' column. If both male and female are NULL, no recoding is performed.}
+
+\item{code_female}{The code used to represent females in the 'sex' column. If both male and female are NULL, no recoding is performed.}
 }
 \value{
 A modified version of the input data.frame \code{ped}, containing an additional or modified 'sex_recode' column where the 'sex' values are recoded according to \code{code_male}. NA values in the 'sex' column are preserved.
