### 2024-02-12 7ef4eb8
finished helper checks
