### 2024-02-12 8e23d52
changed all 1:length
changed from 1:length to seq_along and 1:nrow to seq_len(nrow
