+++ R/simulatePedigree.R
@@ -127,13 +127,13 @@ simulatePedigree <- function(kpc = 3,
       UsedId <- c(UsedIdFemale, UsedIdMale)
 
       # Create spouses
-      for (j in 1:nrow(df_Ngen)) {
+      for (j in seq_len(nrow(df_Ngen))) {
         if (df_Ngen$id[j] %in% UsedId) {
           next
         } else {
           # idx <- j+1
           if (df_Ngen$sex[j] == "F") {
-            for (k in 1:nrow(df_Ngen)) {
+            for (k in seq_len(nrow(df_Ngen))) {
               idr <- df_Ngen$id[k]
               tgt <- (!(idr %in% UsedId)) & df_Ngen$sex[k] == "M"
               # tgt <- ifelse(is.na(tgt),FALSE,TRUE)
@@ -147,7 +147,7 @@ simulatePedigree <- function(kpc = 3,
               }
             }
           } else {
-            for (k in 1:nrow(df_Ngen)) {
+            for (k in seq_len(nrow(df_Ngen))) {
               idr <- df_Ngen$id[k]
               tgt <- (!(idr %in% UsedId)) & df_Ngen$sex[k] == "F"
               # tgt <- ifelse(is.na(tgt),FALSE,TRUE)
@@ -350,7 +350,7 @@ simulatePedigree <- function(kpc = 3,
               UsedId <- c(UsedId, df_Ngen$id[l], df_Ngen$spt[l])
               IdMa <- c(IdMa, rep(df_Ngen$id[l], kpc))
               IdPa <- c(IdPa, rep(df_Ngen$spt[l], kpc))
-            } else if (df_Ngen$ifparent[l] == TRUE & df_Ngen$sex[l] == "M") {
+            } else if (df_Ngen$ifparent[l] == TRUE && df_Ngen$sex[l] == "M") {
               UsedId <- c(UsedId, df_Ngen$id[l], df_Ngen$spt[l])
               IdPa <- c(IdPa, rep(df_Ngen$id[l], kpc))
               IdMa <- c(IdMa, rep(df_Ngen$spt[l], kpc))
@@ -391,7 +391,7 @@ simulatePedigree <- function(kpc = 3,
         # print(IdOfp)
 
         # put the IdMa and IdPa into the dfFam with correspondent OfpId
-        for (m in 1:length(IdOfp)) {
+        for (m in seq_along(IdOfp)) {
           df_Ngen[df_Ngen$id == IdOfp[m], "pat"] <- IdPa[m]
           df_Ngen[df_Ngen$id == IdOfp[m], "mat"] <- IdMa[m]
         }
