+++ R/helper.R
@@ -82,7 +82,7 @@ Null <- function(M) {
 #' simulatePedigree(...)
 #' }
 #' @export
-SimPed <- function(...) {
+SimPed <- function(...) { # nolint: object_name_linter.
   warning("The 'SimPed' function is deprecated. Please use 'simulatePedigree' instead.")
   simulatePedigree(...)
 }
