+++ R/helper.R
@@ -64,6 +64,32 @@ Null <- function(M) {
   return(qr.Q(tmp, complete = TRUE)[, set, drop = FALSE])
 }
 
+
+#' Resample Elements of a Vector
+#'
+#' This function performs resampling of the elements in a vector `x`. It randomly
+#' shuffles the elements of `x` and returns a vector of the resampled elements. If `x`
+#' is empty, it returns `NA_integer_`.
+#'
+#' @param x A vector containing the elements to be resampled. If `x` is empty, the
+#' function will return `NA_integer_`.
+#' @param ... Additional arguments passed to `sample.int`, such as `size` for the
+#' number of items to sample and `replace` indicating whether sampling should be with
+#' replacement.
+#'
+#' @return A vector of resampled elements from `x`. If `x` is empty, returns
+#' `NA_integer_`. The length and type of the returned vector depend on the input
+#' vector `x` and the additional arguments provided via `...`.
+#'
+#'
+#' @export
+resample = function(x, ...){
+  #print(length(x))
+  if(length(x) == 0) return(NA_integer_)
+  x[sample.int(length(x), ...)]
+}
+
+
 #' SimPed (Deprecated)
 #'
 #' This function is a wrapper around the new `simulatePedigree` function.
@@ -132,3 +158,4 @@ relatedness <- function(...) {
   warning("The 'relatedness' function is deprecated. Please use 'inferRelatedness' instead.")
   inferRelatedness(...)
 }
+
