+++ vignettes/network.Rmd
@@ -58,11 +58,12 @@ plotPedigree(potter, code_male = 1, verbose = TRUE)
 We will use the `potter` pedigree data as an example. For convenience, we've renamed the family ID variable to `oldfam` to avoid confusion with the new family ID variable we will create.
 
 ```{r}
-names(potter)[names(potter) == "famID"] <- "oldfam"
+df_potter<-potter
+names(df_potter)[names(df_potter) == "famID"] <- "oldfam"
 
-ds <- ped2fam(potter, famID = "oldfam", personID = "personID")
+ds <- ped2fam(df_potter, famID = "famID", personID = "personID")
 
-table(ds$famID, ds$fam)
+table(ds$famID, ds$oldfam)
 ```
 
 Because the `potter` data already had a family ID variable, we compare 
@@ -87,8 +88,7 @@ add[1:7, 1:7]
 ```
 
 The entry in the ith row and the jth column gives the relatedness
-between person i and person j. For example, person 1 and person 11 share
-`r add[1,11]` of their nuclear DNA.
+between person i and person j. For example, person 1 (`r potter$name[1]`) shares `r add[1,6]` of their nuclear DNA with person 6 (`r potter$name[6]`), shares `r add[1,2]` of their nuclear DNA with person 2 (`r potter$name[2]`).
 
 ```{r}
 table(add)
@@ -128,8 +128,8 @@ mit[1:7, 1:7]
 table(mit)
 ```
 
-As you can see, some of the family members share mitochrondial DNA, such
-as person 1 and person 3 `r mit[1,3]`, whereas person 1 and person 2 do
+As you can see, some of the family members share mitochondrial DNA, such
+as person 2 and person 3 `r mit[2,3]`, whereas person 1 and person 3 do
 not.
 
 ### Computing relatedness through common nuclear environment
