+++ NAMESPACE
@@ -6,6 +6,7 @@ export(calculateRelatedness)
 export(checkIDs)
 export(checkSex)
 export(comp2vech)
+export(dropLink)
 export(evenInsert)
 export(famSizeCal)
 export(fitComponentModel)
@@ -26,6 +27,7 @@ export(plotPedigree)
 export(related_coef)
 export(relatedness)
 export(repairSex)
+export(resample)
 export(simulatePedigree)
 export(vech)
 import(kinship2)
