### 2024-02-12 ba8239e
harmonized xyl's code
harmonized xyl's code and starting to modularize it
