+++ R/simulatePedigree.R
@@ -35,7 +35,6 @@ simulatePedigree <- function(kpc = 3,
                              balancedSex = TRUE,
                              balancedMar = TRUE,
                              verbose = FALSE) {
-
   # SexRatio: ratio of male over female in the offspring setting; used in the between generation combinations
   SexRatio <- sexR / (1 - sexR)
 
@@ -52,16 +51,13 @@ simulatePedigree <- function(kpc = 3,
     # idGen <- ifelse(i==1,
     #                 paste(i,"-",1:sizeGens[i]),
     #                 paste(i,"-",sizeGens[i-1]:sizeGens[i]))
-    ### For each generation, create a seperate dataframe
-    df_Ngen <- data.frame(
-      fam = rep(paste("fam", 1), sizeGens[i], sep = ""),
-      id = idGen[1:sizeGens[i]],
-      gen = rep(i, sizeGens[i]),
-      pat = rep(NA, sizeGens[i]), # father id
-      mat = rep(NA, sizeGens[i]), # mother id
-      spt = rep(NA, sizeGens[i]), # spouse id
-      sex = rep(NA, sizeGens[i])
+    ### For each generation, create a separate dataframe
+    df_Ngen <- createGenDataFrame(
+      sizeGens = sizeGens,
+      genIndex = i,
+      idGen = idGen
     )
+
     ### Let's deal with the sex in each generation first
 
     if (runif(1) > .5) {
@@ -328,43 +324,43 @@ simulatePedigree <- function(kpc = 3,
 
         if (rd_kpc) {
           # get the number of pairs of mates in the i th generation
-          nMates <- sum(df_Ngen$ifparent)/2
-          #cat("number of mates",nMates, "\n")
+          nMates <- sum(df_Ngen$ifparent) / 2
+          # cat("number of mates",nMates, "\n")
           diff <- nMates + 1
-          while(diff > nMates){
-            random_numbers = rpois(nMates, kpc)
-            #cat("original random numbers", random_numbers, "\n")
-            diff = abs(nMates*kpc - sum(random_numbers))
+          while (diff > nMates) {
+            random_numbers <- stats::rpois(nMates, kpc)
+            # cat("original random numbers", random_numbers, "\n")
+            diff <- abs(nMates * kpc - sum(random_numbers))
           }
 
-          #diff = abs(length(IdOfp) - sum(random_numbers))
+          # diff = abs(length(IdOfp) - sum(random_numbers))
           # make sure the sum of kids per couple is equal to the number of kids in the i th generation
-          if (sum(random_numbers) < nMates*kpc){
+          if (sum(random_numbers) < nMates * kpc) {
             names(random_numbers) <- seq(length(random_numbers))
-            random_numbers = sort(random_numbers)
-            random_numbers[1:diff]= random_numbers[1:diff] + 1
-            random_numbers = random_numbers[order(names(random_numbers))]
-          } else if (sum(random_numbers) > nMates*kpc) {
+            random_numbers <- sort(random_numbers)
+            random_numbers[1:diff] <- random_numbers[1:diff] + 1
+            random_numbers <- random_numbers[order(names(random_numbers))]
+          } else if (sum(random_numbers) > nMates * kpc) {
             names(random_numbers) <- seq(length(random_numbers))
-            random_numbers = sort(random_numbers, decreasing = TRUE)
-            random_numbers[1:diff]= random_numbers[1:diff] - 1
-            random_numbers = random_numbers[order(names(random_numbers))]
[…]
