+++ man/identifyComponentModel.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/identification.R
+% Please edit documentation in R/identifyModel.R
 \name{identifyComponentModel}
 \alias{identifyComponentModel}
 \title{identifyComponentModel
