+++ NAMESPACE
@@ -7,6 +7,7 @@ export(checkIDs)
 export(checkSex)
 export(comp2vech)
 export(createGenDataFrame)
+export(determineSex)
 export(dropLink)
 export(evenInsert)
 export(famSizeCal)
