### 2024-02-13 e9ba1de
rename function, expand check
