+++ R/simulatePedigree.R
@@ -181,7 +181,7 @@ simulatePedigree <- function(kpc = 3,
       df_Ngen$ifparent <- FALSE
       df_Ngen$ifson <- FALSE
       df_Ngen$ifdau <- FALSE
-      df_Ngen$IdCouple <- as.character(NA)
+      df_Ngen$coupleId <- NA_character_
       df_Ngen <- df_Ngen[sample(nrow(df_Ngen)), ]
       # Start to connect children with mother and father
       # Step 2.1: mark a group of potential sons and daughters in the i th generation
@@ -190,23 +190,11 @@ simulatePedigree <- function(kpc = 3,
       # count the number of couples in the i th gen
       countCouple <- (nrow(df_Ngen) - sum(is.na(df_Ngen$spt))) * .5
 
-      # give each member a coupleId
-      UsedCoupleId <- character()
-      for (j in 1:sizeGens[i]) {
-        if (df_Ngen$IdCouple[j] %in% UsedCoupleId) {
-          next
-        } else {
-          if (is.na(df_Ngen$spt[j])) {
-            df_Ngen$IdCouple[j] <- as.character(NA)
-          } else {
-            df_Ngen$IdCouple[j] <- paste(sort(c(df_Ngen$id[j], df_Ngen$spt[j]))[1],
-              sort(c(df_Ngen$id[j], df_Ngen$spt[j]))[2],
-              sep = "_"
-            )
-            UsedCoupleId <- c(UsedCoupleId, df_Ngen$IdCouple[j])
-          }
-        }
-      }
+      
+    # Now, assign couple IDs for the current generation
+     df_Ngen <- assignCoupleIds(df_Ngen)
+
+
       # get the number of linked female and male children after excluding the single children
       # get a vector of single person id in the ith generation
       IdSingle <- df_Ngen$id[is.na(df_Ngen$spt)]
@@ -215,7 +203,7 @@ simulatePedigree <- function(kpc = 3,
       SingleM <- sum(df_Ngen$sex == "M" & is.na(df_Ngen$spt))
       CoupleM <- N_LinkedMale - SingleM
       # get all couple ids
-      coupleID <- unique(df_Ngen$IdCouple[!is.na(df_Ngen$IdCouple)])
+      coupleID <- unique(df_Ngen$coupleId[!is.na(df_Ngen$coupleId)])
       if (i == Ngen) {
         CoupleF <- 0
       }
@@ -233,9 +221,9 @@ simulatePedigree <- function(kpc = 3,
             # UsedId <- c(UsedId, df_Ngen$id[j])
           }
         } else {
-          if (df_Ngen$IdCouple[j] %in% coupleBoy & df_Ngen$sex[j] == "M") {
+          if (df_Ngen$coupleId[j] %in% coupleBoy && df_Ngen$sex[j] == "M") {
             df_Ngen$ifson[j] <- TRUE
-          } else if (df_Ngen$IdCouple[j] %in% coupleGirl & df_Ngen$sex[j] == "F") {
+          } else if (df_Ngen$coupleId[j] %in% coupleGirl && df_Ngen$sex[j] == "F") {
             df_Ngen$ifdau[j] <- TRUE
           } else {
             next
@@ -257,7 +245,7 @@ simulatePedigree <- function(kpc = 3,
       df_Ngen$ifdau <- FALSE
       df_Ngen <- df_Ngen[sample(nrow(df_Ngen)), ]
       # Create a pool for the used parents
-      UsedIdParents <- numeric()
+      usedParentIds <- numeric()
 
       for (k in 1:sizeGens[i - 1]) {
         # first check if the number of married couples surpass the marriage rate
@@ -265,10 +253,10 @@ simulatePedigree <- function(kpc = 3,
           break
         } else {
[…]
