+++ man/makeInbreeding.Rd
@@ -4,7 +4,7 @@
 \alias{makeInbreeding}
 \title{makeInbreeding
 A function to create inbred mates in the simulated pedigree \code{data.frame}.
-Inbred mates can be created by specifying their IDs or by specifying the generation the inbred mate should be created.
+Inbred mates can be created by specifying their IDs or the generation the inbred mate should be created.
 When specifying the generation, inbreeding between siblings or 1st cousin needs to be specified.
 This is a supplementary function for \code{simulatePedigree}.}
 \usage{
@@ -36,7 +36,7 @@ Returns a \code{data.frame} with some inbred mates.
 \description{
 makeInbreeding
 A function to create inbred mates in the simulated pedigree \code{data.frame}.
-Inbred mates can be created by specifying their IDs or by specifying the generation the inbred mate should be created.
+Inbred mates can be created by specifying their IDs or the generation the inbred mate should be created.
 When specifying the generation, inbreeding between siblings or 1st cousin needs to be specified.
 This is a supplementary function for \code{simulatePedigree}.
 }

[…]
