### 2024-02-13 3efc1c6
cyclocomp numbers are finally reasonable
Technically, now simulatePedigree has a complexity of 3...  down from 72 😁 However that's because I split it up into a few stages, and now the two main subfunctions are 32 and 28.
