### 2024-02-13 R-Computing-Lab/BGmisc#13 @smasongarrison
Original...
```
                     name cyclocomp
34       simulatePedigree        72
19                ped2com        35
12         makeInbreeding        28
13              makeTwins        17
3                checkIDs        14
11       inferRelatedness        11
26           plotPedigree        11
21              ped2graph         9
4                checkSex         7
7              evenInsert         6
10 identifyComponentModel         6
36   standardize_colnames         6
2    calculateRelatedness         5
5               comp2vech         5
1                 allGens         4
8              famSizeCal         4
9       fitComponentModel         3
27              recodeSex         3
35            sizeAllGens         3
14                   Null         2
15               nullToNA         2
6                   efunc         1
16                ped2add         1
17                 ped2ce         1
18                 ped2cn         1
20                ped2fam         1
22           ped2maternal         1
23                ped2mit         1
24                 ped2mt         1
25           ped2paternal         1
28           related_coef         1
29            relatedness         1
30              repairIDs         1
31              repairSex         1
32                   rmvn         1
33                 SimPed         1
37                 try_na         1
38                   vech         1
```

```
> cyclocomp::cyclocomp_package_dir()
── R CMD build ───────────────────────────────────────────────────────────────────
✔  checking for file 'C:\Users\smaso\AppData\Local\Temp\RtmpohJ6TU\remotes5a4c6a743b96\BGmisc/DESCRIPTION' ...
─  preparing 'BGmisc': (781ms)
✔  checking vignette meta-information
─  building 'BGmisc_1.2.0.tar.gz'

* installing *source* package 'BGmisc' ...
** using staged installation
** R
** data
*** moving datasets to lazyload DB
** inst
** byte-compile and prepare package for lazy loading
** help
*** installing help indices
** building package indices
** installing vignettes
** testing if installed package can be loaded from temporary location
** testing if installed package can be loaded from final location
** testing if installed package keeps a record of temporary installation path
* DONE (BGmisc)
                      name cyclocomp
27                 ped2com        35
4  buildBetweenGenerations        32
5   buildWithinGenerations        28
19          makeInbreeding        28
20               makeTwins        17
21   markPotentialChildren        16
7                 checkIDs        14
18        inferRelatedness        11
34            plotPedigree        11
3          assignCoupleIds        10
8                 checkSex         9
29               ped2graph         9
1      adjustKidsPerCouple         7
14              evenInsert         6
17  identifyComponentModel         6
35               recodeSex         6
45     standardizeColnames         6
6     calculateRelatedness         5
9                comp2vech         5
2                  allGens         4
12                dropLink         4
15              famSizeCal         4
16       fitComponentModel         3
40                resample         3
43        simulatePedigree         3
44             sizeAllGens         3
11            determineSex         2
22                    Null         2
23                nullToNA         2
10      createGenDataFrame         1
13                   efunc         1
24                 ped2add         1
25                  ped2ce         1
26                  ped2cn         1
28                 ped2fam         1
30            ped2maternal         1
31                 ped2mit         1
32                  ped2mt         1
33            ped2paternal         1
36            related_coef         1
37             relatedness         1
38               repairIDs         1
39               repairSex         1
41                    rmvn         1
42                  SimPed         1
46                  try_na         1
47                    vech         1
``` #
