### 2024-02-13 R-Computing-Lab/BGmisc#13 @smasongarrison
I _still_ need to implement sapply, but everything else is pretty solid. (I really appreciate learning about covr and good practice those are awesome tools!)
