### 2024-02-15 255a727
added extra test AND removed last sapply
