+++ R/family.R
@@ -111,8 +111,10 @@ ped2graph <- function(ped,
       )
     )
     edges <- rbind(
-      as.matrix(sapply(ped[, c(personID, momID)], as.character)),
-      as.matrix(sapply(ped[, c(personID, dadID)], as.character))
+      as.matrix(data.frame(personID = as.character(ped[[personID]]),
+                           momID = as.character(ped[[momID]]))),
+      as.matrix(data.frame(personID = as.character(ped[[personID]]),
+                           dadID = as.character(ped[[dadID]])))
     )
   } else if (adjacent == "mothers") {
     nodes <- unique(
@@ -120,14 +122,16 @@ ped2graph <- function(ped,
         as.character(c(ped[[personID]], ped[[momID]]))
       )
     )
-    edges <- as.matrix(sapply(ped[, c(personID, momID)], as.character))
+    edges <- as.matrix(data.frame(personID = as.character(ped[[personID]]),
+                                  momID = as.character(ped[[momID]])))
   } else if (adjacent == "fathers") {
     nodes <- unique(
       stats::na.omit(
         as.character(c(ped[[personID]], ped[[dadID]]))
       )
     )
-    edges <- as.matrix(sapply(ped[, c(personID, dadID)], as.character))
+    edges <-as.matrix(data.frame(personID = as.character(ped[[personID]]),
+                                 dadID = as.character(ped[[dadID]])))
   }
   edges <- edges[stats::complete.cases(edges), ]
 
