### 2024-02-15 R-Computing-Lab/BGmisc#12 @smasongarrison
Thanks again for introducing me to covr. It's a really neat tool.

I'm pretty pleased with how high I was able to get the coverage. At this point, most of the things not covered are for handling specific kinds of data sets, file structures, or specific errors. (Such as someone trying to fit a component model using fitComponentModel that's not identified. I'll eventually task a student with implementing a check to confirm that that error message triggers)

```
> covr::package_coverage()
BGmisc Coverage: 85.36%
R/identifyModel.R: 62.96%
R/checkSex.R: 69.44%
R/tweakPedigree.R: 70.27%
R/helpPedigree.R: 75.00%
R/evenInsert.R: 80.00%
R/helper.R: 81.48%
R/famSizeCal.R: 85.29%
R/family.R: 88.73%
R/convertPedigree.R: 92.59%
R/plotPedigree.R: 93.33%
R/checkIDs.R: 94.74%
R/cleanPedigree.R: 95.45%
R/simulatePedigree.R: 99.43%
R/formula.R: 100.00%
```
