+++ NEWS.md
@@ -1,5 +1,8 @@
+# BGmisc 1.2.0
+* Added numerous code checks, replaced sapply usage, and added additional data validation checks.
+* 
 # BGmisc 1.1.0
-* Added ability to simulate twins, trace paternal and maternal lines, and Harry Potter pedigree
+* Added ability to simulate twins, trace paternal and maternal lines, and a Harry Potter pedigree
 
 # BGmisc 1.0.1
 * Hot fix to resolve plotPedigree wrapper function breaking for pedigrees that contained multiple families
