### 2024-02-16 ee8b4a3
Update NEWS.md
