+++ vignettes/articles/paper.md
@@ -1,6 +1,8 @@
 ---
 title: 'BGmisc: An R Package for Extended Behavior Genetics Analysis'
 output:
+  rmarkdown::html_vignette:
+    keep_md: TRUE
   rticles::joss_article:
     keep_md: TRUE
 tags:
@@ -30,23 +32,21 @@ affiliations:
    index: 2
  - name: Institute for Behavioral Genetics, University of Colorado at Boulder, Boulder, USA 
    index: 3
- - name: Department of Psychology & Neuroscience, University of Colorado at Boulder, Boulder, USA
+ - name: Department of Psychology \& Neuroscience, University of Colorado at Boulder, Boulder, USA
    index: 4
  - name: Department of Psychology, Michigan State University, Michigan, USA
    index: 5
 citation_author: Garrison, Hunter, Lyu, Trattner, & Burt
 csl: apa.csl
 journal: JOSS
-date: "28 September, 2023"
+date: "16 February, 2024"
 bibliography: paper.bib
 vignette: >
   %\VignetteEncoding{UTF-8}
   %\VignetteIndexEntry{modelingandrelatedness}
   %\VignetteIndexEntry{pedigree}
   %\VignetteEngine{knitr::rmarkdown}
- 
 ---
-
 
 
 
@@ -66,14 +66,19 @@ Acknowledgment of any financial support.
 
 <!--  A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience. -->
 
-The field of behavior genetics focuses on illuminating genetic and environmental influences on individual differences.
-Traditionally, twin studies have been at the forefront of this discipline. However, research has moved beyond the classical twin design to embrace more complex family structures such as children of twins (CoT) [@DOnofrio2003], mother-daughter-aunt-niece (MDAN) [@rodgers_mdan], and other extended family designs. These expansions allow for a deeper, more nuanced exploration of genetic and environmental influences, but it also introduces challenges, particularly in data structuring and modeling. In particular, the data structures inherent in these more complicated family designs are orders of magnitude larger than traditional designs. In the classical twin study, for example, a family will consist of a single pair of twins (i.e., two people), whereas in the MDAN design, a family consists of two mother-daughter pairs (i.e., four people). This problem quickly becomes intractable when applied to very extended family pedigrees, where a single family can be of any size.  The `BGmisc` package addresses this gap by offering a comprehensive suite of functions for structuring and modeling extended family pedigree data.
-
+Behavior genetics is a field that studies how our genes and environment contribute to differences in behavior and traits among individuals.
+Traditionally, twin studies have long been a cornerstone of this field, helping researchers understand how genetics influence behavior.
+Recently, the focus has expanded to include studies with more complex family structures, e.g., children of twins [CoT, @DOnofrio2003], mother-daughter-aunt-niece, [MDAN, @rodgers_mdan].
+These broader studies offer more detailed insights into how our genes and environment shape us, but they also make analyzing and organizing the data more complex.
+`BGmisc` simplifies the analysis of these complex data structures by offering a comprehensive suite of functions for accommodating families of any size, from twins to extensive pedigrees.
 
 # Statement of need
 <!-- A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work. -->
 
-As behavior genetics delves into more complex data structures like extended pedigrees, the limitations of current tools become evident. Understandably, packages like `OpenMx` [@Neale2016], `EasyMx` [@easy], and `kinship2` [@kinship2; @kinship2R]  were built for  smaller families and classical designs. In contrast, the `BGmisc` R package was specifically developed to structure and model extended family pedigree data.
[…]
