+++ vignettes/articles/paper.Rmd
@@ -1,9 +1,9 @@
 ---
 title: 'BGmisc: An R Package for Extended Behavior Genetics Analysis'
 output:
-  rticles::joss_article:
-    keep_md: TRUE
   rmarkdown::html_vignette:
+    keep_md: TRUE
+  rticles::joss_article:
     keep_md: TRUE
 tags:
   - R
@@ -46,13 +46,11 @@ vignette: >
   %\VignetteIndexEntry{modelingandrelatedness}
   %\VignetteIndexEntry{pedigree}
   %\VignetteEngine{knitr::rmarkdown}
-
 ---
 
 ```{r setup, echo=FALSE}
 options(rmarkdown.html_vignette.check_title = FALSE)
 ```
-
 
 <!--Guidance 
 JOSS welcomes submissions from broadly diverse research areas. For this reason, we require that authors include in the paper some sentences that explain the software functionality and domain of use to a non-specialist reader. We also require that authors explain the research applications of the software. The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.
@@ -70,23 +68,19 @@ Acknowledgment of any financial support.
 
 <!--  A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience. -->
 
-Behavior genetics explores how our genes and environment shape our individual differences. 
-
-Although twin studies have long been a cornerstone of this field, the `BGmisc` R package introduces advanced tools for analyzing more complex family structures (e.g., children of twins (CoT) [@DOnofrio2003], mother-daughter-aunt-niece (MDAN) [@rodgers_mdan]). 
-This expansions allows for a deeper, more nuanced exploration of genetic and environmental influences, accommodating families of any size, from twins to extensive pedigrees. 
-`BGmisc` simplifies the analysis of these complex data structures by offering a comprehensive suite of functions for structuring and modeling extended family pedigree data.
-
+Behavior genetics is a field that studies how our genes and environment contribute to differences in behavior and traits among individuals.
+Traditionally, twin studies have long been a cornerstone of this field, helping researchers understand how genetics influence behavior.
+Recently, the focus has expanded to include studies with more complex family structures, e.g., children of twins [CoT, @DOnofrio2003], mother-daughter-aunt-niece, [MDAN, @rodgers_mdan].
+These broader studies offer more detailed insights into how our genes and environment shape us, but they also make analyzing and organizing the data more complex.
+`BGmisc` simplifies the analysis of these complex data structures by offering a comprehensive suite of functions for accommodating families of any size, from twins to extensive pedigrees.
 
 # Statement of need
 <!-- A Statement of need section that clearly illustrates the research purpose of the software and places it in the context of related work. -->
 
-The move towards analyzing complex family structures in behavior genetics introduces challenges, particularly in data structuring and modeling. The data structures inherent in these more complicated family designs are orders of magnitude larger than traditional designs. 
-In the classical twin study, for example, a family will consist of a single pair of twins (i.e., two people), whereas in the MDAN design, 
-a family consists of two mother-daughter pairs (i.e., four people). This problem quickly becomes intractable when applied to very extended family pedigrees, where a single family can be of any size.  The `BGmisc` package addresses this gap by offering a comprehensive suite of functions for structuring and modeling extended family pedigree data.
+The move towards analyzing complex family structures in behavior genetics introduces challenges in data structuring and modeling. The data structures inherent in these more complicated family designs are orders of magnitude larger than traditional designs.
[…]
