### 2024-02-16 #6203 @smasongarrison
@osorensen I've now addressed all the issues raised by the reviewers.
- [X] Install instructions for cran are now present #https://github.com/R-Computing-Lab/BGmisc/pull/11
- [X] test coverage has been increased from 62% to 85% #https://github.com/R-Computing-Lab/BGmisc/issues/12
- [X] code complexity has been reduced to a reasonable range (most complex function is 34, instead of 71) #https://github.com/R-Computing-Lab/BGmisc/issues/13
- [X] made the text more accessible to people outside the field of behavior genetics #https://github.com/R-Computing-Lab/BGmisc/issues/20
