### 2024-02-17 2fb3f42
Update paper.bib
added doi for the missing bga cites (and yep, they're all the same because of how the journal works for published abstracts)
