+++ vignettes/articles/paper.bib
@@ -109,6 +109,7 @@
     journal = {Behavior Genetics},
     year = {2023},
     month = {June},
+    doi = {10.1007/s10519-023-10156-9},
     note = {Abstract presented at the Behavior Genetics Annual Meeting, Murcia, Spain}
 }
 @article{burt_mom_genes_2023,
@@ -117,6 +118,7 @@
     journal = {Behavior Genetics},
     year = {2023},
     month = {June},
+    doi = {10.1007/s10519-023-10156-9},
     note = {Abstract presented at the 53rd Behavior Genetics Association Annual Meeting, Murcia, Spain}
 }
 @mastersthesis{lyu_masters_thesis_2023,
@@ -165,6 +167,7 @@
     journal = {Behavior Genetics},
     year = {2023},
     month = {June},
+    doi = {10.1007/s10519-023-10156-9},
     note = {Abstract presented at the 53rd Behavior Genetics Association Annual Meeting, Murcia, Spain}
 }
 @article{lyu_statistical_power_2023,
@@ -173,6 +176,7 @@
     journal = {Behavior Genetics},
     year = {2023},
     month = {June},
+    doi = {10.1007/s10519-023-10156-9},
     note = {Abstract presented at the 53rd Behavior Genetics Association Annual Meeting, Murcia, Spain}
 }
 
