### 2024-02-17 6b05309
Update paper.bib
added doi for the missing bga cites (and yep, they're all the same because of how the journal works for published abstracts)

Also added acknowledge for Rachel Good

Update paper.Rmd

added additional acknowledgement
