### 2024-02-17 #6203 @smasongarrison
Version number in the release is 1.2.0
