+++ vignettes/articles/paper.bib
@@ -51,7 +51,7 @@
 	issn = {0033-3123},
 	shorttitle = {{OpenMx} 2.0},
 	url = {http://link.springer.com/10.1007/s11336-014-9435-8},
-	doi = {10/f8rfrg},
+	doi = {10.1007/s11336-014-9435-8},
 	abstract = {Psychometrika, doi:10.1007/s11336-014-9435-8},
 	language = {en},
 	number = {2},
