+++ vignettes/articles/paper.bib
@@ -135,7 +135,7 @@
 	volume = {44},
 	issn = {0021-9630},
 	url = {http://www.ncbi.nlm.nih.gov/pubmed/14626455},
-	doi = {10/ff3dbm},
+	doi = {10.1111/1469-7610.00196},
 	language = {eng},
 	number = {8},
 	urldate = {2016-10-03},
