+++ README.md
@@ -12,7 +12,8 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 version](https://www.r-pkg.org/badges/version/BGmisc)](https://cran.r-project.org/package=BGmisc)
 [![Package
 downloads](https://cranlogs.r-pkg.org/badges/grand-total/BGmisc)](https://cran.r-project.org/package=BGmisc)</br>
-[![.github/workflows/draft-pdf.yml](https://github.com/R-Computing-Lab/BGmisc/actions/workflows/draft-pdf.yml/badge.svg)](https://github.com/R-Computing-Lab/BGmisc/actions/workflows/draft-pdf.yml)
+[![status](https://joss.theoj.org/papers/ee3a025be4f61584f977a7657d936187/status.svg)](https://joss.theoj.org/papers/ee3a025be4f61584f977a7657d936187)
+
 [![R-CMD-check](https://github.com/R-Computing-Lab/BGmisc/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/R-Computing-Lab/BGmisc/actions/workflows/R-CMD-check.yaml)
 [![Dev Main
 branch](https://github.com/R-Computing-Lab/BGmisc/actions/workflows/R-CMD-dev_maincheck.yaml/badge.svg)](https://github.com/R-Computing-Lab/BGmisc/actions/workflows/R-CMD-dev_maincheck.yaml)
@@ -39,6 +40,34 @@ To install the development version of BGmisc from
 ``` r
 # install.packages("devtools")
 devtools::install_github("R-Computing-Lab/BGmisc")
+```
+
+## Citation
+
+If you use BGmisc in your research or wish to refer to it, please cite
+the following paper:
+
+``` r
+citation(package="BGmisc")
+To cite BGmisc in publications use:
+
+  Garrison, S. Mason, Hunter, Michael D., Lyu, Xuanyu, Trattner,
+  Jonathan D., Burt, S. Alexandra (2024). "BGmisc: An R Package for
+  Extended Behavior Genetics Analysis." _Journal of Open Source
+  Software_, *9*(94). doi:10.21105/joss.06203
+  <https://doi.org/10.21105/joss.06203>.
+
+A BibTeX entry for LaTeX users is
+
+  @Article{bgmisc,
+    title = {BGmisc: An R Package for Extended Behavior Genetics Analysis},
+    author = {{Garrison, S. Mason} and {Hunter, Michael D.} and {Lyu, Xuanyu} and {Trattner, Jonathan D.} and {Burt, S. Alexandra}},
+    journal = {Journal of Open Source Software},
+    year = {2024},
+    volume = {9},
+    number = {94},
+    doi = {10.21105/joss.06203},
+  }
 ```
 
 ## Contributing
