+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://r-computing-lab.github.io/BGmisc/
+template:
+  bootstrap: 5
+
