+++ pkgdown.yml
@@ -1,0 +1,14 @@
+pandoc: 2.19.2
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles:
+  analyticrelatedness: analyticrelatedness.html
+  paper: paper.html
+  modelingrelatedness: modelingrelatedness.html
+  network: network.html
+  pedigree: pedigree.html
+last_built: 2024-02-22T17:39Z
+urls:
+  reference: https://r-computing-lab.github.io/BGmisc/reference
+  article: https://r-computing-lab.github.io/BGmisc/articles
+
