### 2024-02-23 378ad9d
prepping for rcran update
Create CITATION.cff

added skips for verbose checks
