+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   modelingrelatedness: modelingrelatedness.html
   network: network.html
   pedigree: pedigree.html
-last_built: 2024-02-22T17:39Z
+last_built: 2024-02-23T16:25Z
 urls:
   reference: https://r-computing-lab.github.io/BGmisc/reference
   article: https://r-computing-lab.github.io/BGmisc/articles
