### 2024-02-24 #6203 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
