### 2024-02-24 #6203 @oliviaguest
Huge thanks to the reviewers @wjakethompson, @mhu48; and the editor @osorensen! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to @smasongarrison! 🥳 🍾
